Crate google_displayvideo1

This documentation was generated from Display Video crate version 2.0.0+20210325, where 20210325 is the exact revision of the displayvideo:v1 schema built by the mako code generator v2.0.0.

Everything else about the Display Video v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

advertisers

assets upload, audit, bulk edit advertiser assigned targeting options, bulk list advertiser assigned targeting options, campaigns bulk list campaign assigned targeting options, campaigns create, campaigns delete, campaigns get, campaigns list, campaigns patch, campaigns targeting types assigned targeting options get, campaigns targeting types assigned targeting options list, channels create, channels get, channels list, channels patch, channels sites bulk edit, channels sites create, channels sites delete, channels sites list, channels sites replace, create, creatives create, creatives delete, creatives get, creatives list, creatives patch, delete, get, insertion orders bulk list insertion order assigned targeting options, insertion orders create, insertion orders delete, insertion orders get, insertion orders list, insertion orders patch, insertion orders targeting types assigned targeting options get, insertion orders targeting types assigned targeting options list, line items bulk edit line item assigned targeting options, line items bulk list line item assigned targeting options, line items create, line items delete, line items generate default, line items get, line items list, line items patch, line items targeting types assigned targeting options create, line items targeting types assigned targeting options delete, line items targeting types assigned targeting options get, line items targeting types assigned targeting options list, list, location lists assigned locations bulk edit, location lists assigned locations create, location lists assigned locations delete, location lists assigned locations list, location lists create, location lists get, location lists list, location lists patch, manual triggers activate, manual triggers create, manual triggers deactivate, manual triggers get, manual triggers list, manual triggers patch, negative keyword lists create, negative keyword lists delete, negative keyword lists get, negative keyword lists list, negative keyword lists negative keywords bulk edit, negative keyword lists negative keywords create, negative keyword lists negative keywords delete, negative keyword lists negative keywords list, negative keyword lists negative keywords replace, negative keyword lists patch, patch, targeting types assigned targeting options create, targeting types assigned targeting options delete, targeting types assigned targeting options get and targeting types assigned targeting options list

combined audiences

get and list

custom bidding algorithms

get and list

custom lists

get and list

first and third party audiences

get and list

floodlight groups

get and patch

google audiences

get and list

inventory source groups

assigned inventory sources bulk edit, assigned inventory sources create, assigned inventory sources delete, assigned inventory sources list, create, delete, get, list and patch

inventory sources

get and list

media

download

partners

bulk edit partner assigned targeting options, channels create, channels get, channels list, channels patch, channels sites bulk edit, channels sites create, channels sites delete, channels sites list, channels sites replace, get, list, targeting types assigned targeting options create, targeting types assigned targeting options delete, targeting types assigned targeting options get and targeting types assigned targeting options list

sdfdownloadtasks

create and operations get

targeting types

targeting options get, targeting options list and targeting options search

users

bulk edit assigned user roles, create, delete, get, list and patch

Upload supported by ...

assets upload advertisers

Download supported by ...

download media

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit().await

Or specifically ...

let r = hub.advertisers().assets_upload(...).doit().await
let r = hub.advertisers().campaigns_targeting_types_assigned_targeting_options_get(...).doit().await
let r = hub.advertisers().campaigns_targeting_types_assigned_targeting_options_list(...).doit().await
let r = hub.advertisers().campaigns_bulk_list_campaign_assigned_targeting_options(...).doit().await
let r = hub.advertisers().campaigns_create(...).doit().await
let r = hub.advertisers().campaigns_delete(...).doit().await
let r = hub.advertisers().campaigns_get(...).doit().await
let r = hub.advertisers().campaigns_list(...).doit().await
let r = hub.advertisers().campaigns_patch(...).doit().await
let r = hub.advertisers().channels_sites_bulk_edit(...).doit().await
let r = hub.advertisers().channels_sites_create(...).doit().await
let r = hub.advertisers().channels_sites_delete(...).doit().await
let r = hub.advertisers().channels_sites_list(...).doit().await
let r = hub.advertisers().channels_sites_replace(...).doit().await
let r = hub.advertisers().channels_create(...).doit().await
let r = hub.advertisers().channels_get(...).doit().await
let r = hub.advertisers().channels_list(...).doit().await
let r = hub.advertisers().channels_patch(...).doit().await
let r = hub.advertisers().creatives_create(...).doit().await
let r = hub.advertisers().creatives_delete(...).doit().await
let r = hub.advertisers().creatives_get(...).doit().await
let r = hub.advertisers().creatives_list(...).doit().await
let r = hub.advertisers().creatives_patch(...).doit().await
let r = hub.advertisers().insertion_orders_targeting_types_assigned_targeting_options_get(...).doit().await
let r = hub.advertisers().insertion_orders_targeting_types_assigned_targeting_options_list(...).doit().await
let r = hub.advertisers().insertion_orders_bulk_list_insertion_order_assigned_targeting_options(...).doit().await
let r = hub.advertisers().insertion_orders_create(...).doit().await
let r = hub.advertisers().insertion_orders_delete(...).doit().await
let r = hub.advertisers().insertion_orders_get(...).doit().await
let r = hub.advertisers().insertion_orders_list(...).doit().await
let r = hub.advertisers().insertion_orders_patch(...).doit().await
let r = hub.advertisers().line_items_targeting_types_assigned_targeting_options_create(...).doit().await
let r = hub.advertisers().line_items_targeting_types_assigned_targeting_options_delete(...).doit().await
let r = hub.advertisers().line_items_targeting_types_assigned_targeting_options_get(...).doit().await
let r = hub.advertisers().line_items_targeting_types_assigned_targeting_options_list(...).doit().await
let r = hub.advertisers().line_items_bulk_edit_line_item_assigned_targeting_options(...).doit().await
let r = hub.advertisers().line_items_bulk_list_line_item_assigned_targeting_options(...).doit().await
let r = hub.advertisers().line_items_create(...).doit().await
let r = hub.advertisers().line_items_delete(...).doit().await
let r = hub.advertisers().line_items_generate_default(...).doit().await
let r = hub.advertisers().line_items_get(...).doit().await
let r = hub.advertisers().line_items_list(...).doit().await
let r = hub.advertisers().line_items_patch(...).doit().await
let r = hub.advertisers().location_lists_assigned_locations_bulk_edit(...).doit().await
let r = hub.advertisers().location_lists_assigned_locations_create(...).doit().await
let r = hub.advertisers().location_lists_assigned_locations_delete(...).doit().await
let r = hub.advertisers().location_lists_assigned_locations_list(...).doit().await
let r = hub.advertisers().location_lists_create(...).doit().await
let r = hub.advertisers().location_lists_get(...).doit().await
let r = hub.advertisers().location_lists_list(...).doit().await
let r = hub.advertisers().location_lists_patch(...).doit().await
let r = hub.advertisers().manual_triggers_activate(...).doit().await
let r = hub.advertisers().manual_triggers_create(...).doit().await
let r = hub.advertisers().manual_triggers_deactivate(...).doit().await
let r = hub.advertisers().manual_triggers_get(...).doit().await
let r = hub.advertisers().manual_triggers_list(...).doit().await
let r = hub.advertisers().manual_triggers_patch(...).doit().await
let r = hub.advertisers().negative_keyword_lists_negative_keywords_bulk_edit(...).doit().await
let r = hub.advertisers().negative_keyword_lists_negative_keywords_create(...).doit().await
let r = hub.advertisers().negative_keyword_lists_negative_keywords_delete(...).doit().await
let r = hub.advertisers().negative_keyword_lists_negative_keywords_list(...).doit().await
let r = hub.advertisers().negative_keyword_lists_negative_keywords_replace(...).doit().await
let r = hub.advertisers().negative_keyword_lists_create(...).doit().await
let r = hub.advertisers().negative_keyword_lists_delete(...).doit().await
let r = hub.advertisers().negative_keyword_lists_get(...).doit().await
let r = hub.advertisers().negative_keyword_lists_list(...).doit().await
let r = hub.advertisers().negative_keyword_lists_patch(...).doit().await
let r = hub.advertisers().targeting_types_assigned_targeting_options_create(...).doit().await
let r = hub.advertisers().targeting_types_assigned_targeting_options_delete(...).doit().await
let r = hub.advertisers().targeting_types_assigned_targeting_options_get(...).doit().await
let r = hub.advertisers().targeting_types_assigned_targeting_options_list(...).doit().await
let r = hub.advertisers().audit(...).doit().await
let r = hub.advertisers().bulk_edit_advertiser_assigned_targeting_options(...).doit().await
let r = hub.advertisers().bulk_list_advertiser_assigned_targeting_options(...).doit().await
let r = hub.advertisers().create(...).doit().await
let r = hub.advertisers().delete(...).doit().await
let r = hub.advertisers().get(...).doit().await
let r = hub.advertisers().list(...).doit().await
let r = hub.advertisers().patch(...).doit().await

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-displayvideo1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.14"
hyper-rustls = "^0.22"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^5.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_displayvideo1 as displayvideo1;
use displayvideo1::{Result, Error};
use std::default::Default;
use oauth2;
use displayvideo1::DisplayVideo;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = DisplayVideo::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.advertisers().campaigns_targeting_types_assigned_targeting_options_list("advertiserId", "campaignId", "targetingType")
             .page_token("amet.")
             .page_size(-20)
             .order_by("ipsum")
             .filter("gubergren")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be encodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Re-exports

pub use api::DisplayVideo;

pub use client::Result;

pub use client::Error;

pub use client::Delegate;

Modules

api

client