Module google_displayvideo1::api

Structs

ActivateManualTriggerRequest
Request message for ManualTriggerService.ActivateManualTrigger.

ActiveViewVideoViewabilityMetricConfig
Configuration for custom Active View video viewability metrics.

Adloox
Details of Adloox settings.

Advertiser
A single advertiser in Display & Video 360 (DV360).

AdvertiserAdServerConfig
Ad server related settings of an advertiser.

AdvertiserAssetUploadCall
Uploads an asset. Returns the ID of the newly uploaded asset if successful. The asset file size should be no more than 10 MB for images, 200 MB for ZIP files, and 1 GB for videos.

AdvertiserAuditCall
Audits an advertiser. Returns the counts of used entities per resource type under the advertiser provided. Used entities count towards their respective resource limit. See https://support.google.com/displayvideo/answer/6071450.

AdvertiserBulkEditAdvertiserAssignedTargetingOptionCall
Bulk edits targeting options under a single advertiser. The operation will delete the assigned targeting options provided in BulkEditAdvertiserAssignedTargetingOptionsRequest.delete_requests and then create the assigned targeting options provided in BulkEditAdvertiserAssignedTargetingOptionsRequest.create_requests .

AdvertiserBulkListAdvertiserAssignedTargetingOptionCall
Lists assigned targeting options of an advertiser across targeting types.

AdvertiserCampaignBulkListCampaignAssignedTargetingOptionCall
Lists assigned targeting options of a campaign across targeting types.

AdvertiserCampaignCreateCall
Creates a new campaign. Returns the newly created campaign if successful.

AdvertiserCampaignDeleteCall
Permanently deletes a campaign. A deleted campaign cannot be recovered. The campaign should be archived first, i.e. set entity_status to ENTITY_STATUS_ARCHIVED, to be able to delete it.

AdvertiserCampaignGetCall
Gets a campaign.

AdvertiserCampaignListCall
Lists campaigns in an advertiser. The order is defined by the order_by parameter. If a filter by entity_status is not specified, campaigns with ENTITY_STATUS_ARCHIVED will not be included in the results.

AdvertiserCampaignPatchCall
Updates an existing campaign. Returns the updated campaign if successful.

AdvertiserCampaignTargetingTypeAssignedTargetingOptionGetCall
Gets a single targeting option assigned to a campaign.

AdvertiserCampaignTargetingTypeAssignedTargetingOptionListCall
Lists the targeting options assigned to a campaign for a specified targeting type.

AdvertiserChannelCreateCall
Creates a new channel. Returns the newly created channel if successful.

AdvertiserChannelGetCall
Gets a channel for a partner or advertiser.

AdvertiserChannelListCall
Lists channels for a partner or advertiser.

AdvertiserChannelPatchCall
Updates a channel. Returns the updated channel if successful.

AdvertiserChannelSiteBulkEditCall
Bulk edits sites under a single channel. The operation will delete the sites provided in BulkEditSitesRequest.deleted_sites and then create the sites provided in BulkEditSitesRequest.created_sites.

AdvertiserChannelSiteCreateCall
Creates a site in a channel.

AdvertiserChannelSiteDeleteCall
Deletes a site from a channel.

AdvertiserChannelSiteListCall
Lists sites in a channel.

AdvertiserChannelSiteReplaceCall
Replaces all of the sites under a single channel. The operation will replace the sites under a channel with the sites provided in ReplaceSitesRequest.new_sites.

AdvertiserCreateCall
Creates a new advertiser. Returns the newly created advertiser if successful. This method can take up to 180 seconds to complete.

AdvertiserCreativeConfig
Creatives related settings of an advertiser.

AdvertiserCreativeCreateCall
Creates a new creative. Returns the newly created creative if successful.

AdvertiserCreativeDeleteCall
Deletes a creative. Returns error code NOT_FOUND if the creative does not exist. The creative should be archived first, i.e. set entity_status to ENTITY_STATUS_ARCHIVED, before it can be deleted.

AdvertiserCreativeGetCall
Gets a creative.

AdvertiserCreativeListCall
Lists creatives in an advertiser. The order is defined by the order_by parameter. If a filter by entity_status is not specified, creatives with ENTITY_STATUS_ARCHIVED will not be included in the results.

AdvertiserCreativePatchCall
Updates an existing creative. Returns the updated creative if successful.

AdvertiserDataAccessConfig
Settings that control how advertiser related data may be accessed.

AdvertiserDeleteCall
Deletes an advertiser. Deleting an advertiser will delete all of its child resources, for example, campaigns, insertion orders and line items. A deleted advertiser cannot be recovered.

AdvertiserGeneralConfig
General settings of an advertiser.

AdvertiserGetCall
Gets an advertiser.

AdvertiserInsertionOrderBulkListInsertionOrderAssignedTargetingOptionCall
Lists assigned targeting options of an insertion order across targeting types.

AdvertiserInsertionOrderCreateCall
Creates a new insertion order. Returns the newly created insertion order if successful.

AdvertiserInsertionOrderDeleteCall
Deletes an insertion order. Returns error code NOT_FOUND if the insertion order does not exist. The insertion order should be archived first, i.e. set entity_status to ENTITY_STATUS_ARCHIVED, to be able to delete it.

AdvertiserInsertionOrderGetCall
Gets an insertion order. Returns error code NOT_FOUND if the insertion order does not exist.

AdvertiserInsertionOrderListCall
Lists insertion orders in an advertiser. The order is defined by the order_by parameter. If a filter by entity_status is not specified, insertion orders with ENTITY_STATUS_ARCHIVED will not be included in the results.

AdvertiserInsertionOrderPatchCall
Updates an existing insertion order. Returns the updated insertion order if successful.

AdvertiserInsertionOrderTargetingTypeAssignedTargetingOptionGetCall
Gets a single targeting option assigned to an insertion order.

AdvertiserInsertionOrderTargetingTypeAssignedTargetingOptionListCall
Lists the targeting options assigned to an insertion order.

AdvertiserLineItemBulkEditLineItemAssignedTargetingOptionCall
Bulk edits targeting options under a single line item. The operation will delete the assigned targeting options provided in BulkEditLineItemAssignedTargetingOptionsRequest.delete_requests and then create the assigned targeting options provided in BulkEditLineItemAssignedTargetingOptionsRequest.create_requests .

AdvertiserLineItemBulkListLineItemAssignedTargetingOptionCall
Lists assigned targeting options of a line item across targeting types.

AdvertiserLineItemCreateCall
Creates a new line item. Returns the newly created line item if successful.

AdvertiserLineItemDeleteCall
Deletes a line item. Returns error code NOT_FOUND if the line item does not exist. The line item should be archived first, i.e. set entity_status to ENTITY_STATUS_ARCHIVED, to be able to delete it.

AdvertiserLineItemGenerateDefaultCall
Creates a new line item with settings (including targeting) inherited from the insertion order. Returns the newly created line item if successful. There are default values based on the three fields: * The insertion order's InsertionOrderType * The insertion order's InsertionOrderAutomationType * The given line_item_type

AdvertiserLineItemGetCall
Gets a line item.

AdvertiserLineItemListCall
Lists line items in an advertiser. The order is defined by the order_by parameter. If a filter by entity_status is not specified, line items with ENTITY_STATUS_ARCHIVED will not be included in the results.

AdvertiserLineItemPatchCall
Updates an existing line item. Returns the updated line item if successful.

AdvertiserLineItemTargetingTypeAssignedTargetingOptionCreateCall
Assigns a targeting option to a line item. Returns the assigned targeting option if successful.

AdvertiserLineItemTargetingTypeAssignedTargetingOptionDeleteCall
Deletes an assigned targeting option from a line item.

AdvertiserLineItemTargetingTypeAssignedTargetingOptionGetCall
Gets a single targeting option assigned to a line item.

AdvertiserLineItemTargetingTypeAssignedTargetingOptionListCall
Lists the targeting options assigned to a line item.

AdvertiserListCall
Lists advertisers that are accessible to the current user. The order is defined by the order_by parameter. A single partner_id is required. Cross-partner listing is not supported.

AdvertiserLocationListAssignedLocationBulkEditCall
Bulk edits multiple assignments between locations and a single location list. The operation will delete the assigned locations provided in BulkEditAssignedLocationsRequest.deleted_assigned_locations and then create the assigned locations provided in BulkEditAssignedLocationsRequest.created_assigned_locations.

AdvertiserLocationListAssignedLocationCreateCall
Creates an assignment between a location and a location list.

AdvertiserLocationListAssignedLocationDeleteCall
Deletes the assignment between a location and a location list.

AdvertiserLocationListAssignedLocationListCall
Lists locations assigned to a location list.

AdvertiserLocationListCreateCall
Creates a new location list. Returns the newly created location list if successful.

AdvertiserLocationListGetCall
Gets a location list.

AdvertiserLocationListListCall
Lists location lists based on a given advertiser id.

AdvertiserLocationListPatchCall
Updates a location list. Returns the updated location list if successful.

AdvertiserManualTriggerActivateCall
Activates a manual trigger. Each activation of the manual trigger must be at least 5 minutes apart, otherwise an error will be returned.

AdvertiserManualTriggerCreateCall
Creates a new manual trigger. Returns the newly created manual trigger if successful.

AdvertiserManualTriggerDeactivateCall
Deactivates a manual trigger.

AdvertiserManualTriggerGetCall
Gets a manual trigger.

AdvertiserManualTriggerListCall
Lists manual triggers that are accessible to the current user for a given advertiser ID. The order is defined by the order_by parameter. A single advertiser_id is required.

AdvertiserManualTriggerPatchCall
Updates a manual trigger. Returns the updated manual trigger if successful.

AdvertiserMethods
A builder providing access to all methods supported on advertiser resources. It is not used directly, but through the DisplayVideo hub.

AdvertiserNegativeKeywordListCreateCall
Creates a new negative keyword list. Returns the newly created negative keyword list if successful.

AdvertiserNegativeKeywordListDeleteCall
Deletes a negative keyword list given an advertiser ID and a negative keyword list ID.

AdvertiserNegativeKeywordListGetCall
Gets a negative keyword list given an advertiser ID and a negative keyword list ID.

AdvertiserNegativeKeywordListListCall
Lists negative keyword lists based on a given advertiser id.

AdvertiserNegativeKeywordListNegativeKeywordBulkEditCall
Bulk edits negative keywords in a single negative keyword list. The operation will delete the negative keywords provided in BulkEditNegativeKeywordsRequest.deleted_negative_keywords and then create the negative keywords provided in BulkEditNegativeKeywordsRequest.created_negative_keywords. This operation is guaranteed to be atomic and will never result in a partial success or partial failure.

AdvertiserNegativeKeywordListNegativeKeywordCreateCall
Creates a negative keyword in a negative keyword list.

AdvertiserNegativeKeywordListNegativeKeywordDeleteCall
Deletes a negative keyword from a negative keyword list.

AdvertiserNegativeKeywordListNegativeKeywordListCall
Lists negative keywords in a negative keyword list.

AdvertiserNegativeKeywordListNegativeKeywordReplaceCall
Replaces all negative keywords in a single negative keyword list. The operation will replace the keywords in a negative keywords with keywords provided in ReplaceNegativeKeywordsRequest.new_negative_keywords.

AdvertiserNegativeKeywordListPatchCall
Updates a negative keyword list. Returns the updated negative keyword list if successful.

AdvertiserPatchCall
Updates an existing advertiser. Returns the updated advertiser if successful.

AdvertiserSdfConfig
Structured Data Files (SDF) settings of an advertiser.

AdvertiserTargetingConfig
Targeting settings related to ad serving of an advertiser.

AdvertiserTargetingTypeAssignedTargetingOptionCreateCall
Assigns a targeting option to an advertiser. Returns the assigned targeting option if successful.

AdvertiserTargetingTypeAssignedTargetingOptionDeleteCall
Deletes an assigned targeting option from an advertiser.

AdvertiserTargetingTypeAssignedTargetingOptionGetCall
Gets a single targeting option assigned to an advertiser.

AdvertiserTargetingTypeAssignedTargetingOptionListCall
Lists the targeting options assigned to an advertiser.

AgeRangeAssignedTargetingOptionDetails
Represents a targetable age range. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_AGE_RANGE.

AgeRangeTargetingOptionDetails
Represents a targetable age range. This will be populated in the age_range_details field when targeting_type is TARGETING_TYPE_AGE_RANGE.

AppAssignedTargetingOptionDetails
Details for assigned app targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_APP.

AppCategoryAssignedTargetingOptionDetails
Details for assigned app category targeting option. This will be populated in the app_category_details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_APP_CATEGORY.

AppCategoryTargetingOptionDetails
Represents a targetable collection of apps. A collection lets you target dynamic groups of related apps that are maintained by the platform, for example All Apps/Google Play/Games. This will be populated in the app_category_details field when targeting_type is TARGETING_TYPE_APP_CATEGORY.

Asset
A single asset.

AssetAssociation
Asset association for the creative.

AssignedInventorySource
An assignment between a targetable inventory source and an inventory source group.

AssignedLocation
An assignment between a location list and a relevant targeting option. Currently, geo region targeting options are the only supported option for assignment.

AssignedTargetingOption
A single assigned targeting option, which defines the state of a targeting option for an entity with targeting settings.

AssignedUserRole
A single assigned user role, which defines a user's authorized interaction with a specified partner or advertiser.

AudienceGroupAssignedTargetingOptionDetails
Assigned audience group targeting option details. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_AUDIENCE_GROUP. The relation between each group is UNION, except for excluded_first_and_third_party_audience_group and excluded_google_audience_group, of which COMPLEMENT is UNION'ed with other groups.

AudioVideoOffset
The length an audio or a video has been played.

AuditAdvertiserResponse
Response message for AdvertiserService.AuditAdvertiser.

AuthorizedSellerStatusAssignedTargetingOptionDetails
Represents an assigned authorized seller status. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_AUTHORIZED_SELLER_STATUS.

AuthorizedSellerStatusTargetingOptionDetails
Represents a targetable authorized seller status. This will be populated in the authorized_seller_status_details field when targeting_type is TARGETING_TYPE_AUTHORIZED_SELLER_STATUS.

BiddingStrategy
Settings that control the bid strategy. Bid strategy determines the bid price.

BrowserAssignedTargetingOptionDetails
Details for assigned browser targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_BROWSER.

BrowserTargetingOptionDetails
Represents a targetable browser. This will be populated in the browser_details field when targeting_type is TARGETING_TYPE_BROWSER.

BulkEditAdvertiserAssignedTargetingOptionsRequest
Request message for BulkEditAdvertiserAssignedTargetingOptions.

BulkEditAdvertiserAssignedTargetingOptionsResponse
There is no detailed description.

BulkEditAssignedInventorySourcesRequest
Request message for AssignedInventorySourceService.BulkEdit.

BulkEditAssignedInventorySourcesResponse
Response message for AssignedInventorySourceService.BulkEdit.

BulkEditAssignedLocationsRequest
Request message for AssignedLocationService.BulkEditAssignedLocations.

BulkEditAssignedLocationsResponse
Response message for AssignedLocationService.BulkEditAssignedLocations.

BulkEditAssignedUserRolesRequest
Request message for BulkEditAssignedUserRoles.

BulkEditAssignedUserRolesResponse
There is no detailed description.

BulkEditLineItemAssignedTargetingOptionsRequest
Request message for BulkEditLineItemAssignedTargetingOptions.

BulkEditLineItemAssignedTargetingOptionsResponse
There is no detailed description.

BulkEditNegativeKeywordsRequest
Request message for NegativeKeywordService.BulkEditNegativeKeywords.

BulkEditNegativeKeywordsResponse
Response message for NegativeKeywordService.BulkEditNegativeKeywords.

BulkEditPartnerAssignedTargetingOptionsRequest
Request message for BulkEditPartnerAssignedTargetingOptions.

BulkEditPartnerAssignedTargetingOptionsResponse
There is no detailed description.

BulkEditSitesRequest
Request message for SiteService.BulkEditSites.

BulkEditSitesResponse
Response message for SiteService.BulkEditSites.

BulkListAdvertiserAssignedTargetingOptionsResponse
There is no detailed description.

BulkListCampaignAssignedTargetingOptionsResponse
Response message for BulkListCampaignAssignedTargetingOptions.

BulkListInsertionOrderAssignedTargetingOptionsResponse
Response message for BulkListInsertionOrderAssignedTargetingOptions.

BulkListLineItemAssignedTargetingOptionsResponse
There is no detailed description.

Campaign
A single campaign.

CampaignFlight
Settings that track the planned spend and duration of a campaign.

CampaignGoal
Settings that control the goal of a campaign.

CarrierAndIspAssignedTargetingOptionDetails
Details for assigned carrier and ISP targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_CARRIER_AND_ISP.

CarrierAndIspTargetingOptionDetails
Represents a targetable carrier or ISP. This will be populated in the carrier_and_isp_details field of a TargetingOption when targeting_type is TARGETING_TYPE_CARRIER_AND_ISP.

CategoryAssignedTargetingOptionDetails
Assigned category targeting option details. This will be populated in the category_details field when targeting_type is TARGETING_TYPE_CATEGORY.

CategoryTargetingOptionDetails
Represents a targetable category. This will be populated in the category_details field of a TargetingOption when targeting_type is TARGETING_TYPE_CATEGORY.

Channel
A single channel. Channels are custom groups of related websites and apps.

ChannelAssignedTargetingOptionDetails
Details for assigned channel targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_CHANNEL.

CmHybridConfig
Settings for advertisers that use both Campaign Manager 360 (CM360) and third-party ad servers.

CmTrackingAd
A Campaign Manager 360 tracking ad.

CombinedAudience
Describes a combined audience resource.

CombinedAudienceGetCall
Gets a combined audience.

CombinedAudienceGroup
Details of combined audience group. All combined audience targeting settings are logically ‘OR’ of each other.

CombinedAudienceListCall
Lists combined audiences. The order is defined by the order_by parameter.

CombinedAudienceMethods
A builder providing access to all methods supported on combinedAudience resources. It is not used directly, but through the DisplayVideo hub.

CombinedAudienceTargetingSetting
Details of combined audience targeting setting.

ContentInstreamPositionAssignedTargetingOptionDetails
Assigned content instream position targeting option details. This will be populated in the content_instream_position_details field when targeting_type is TARGETING_TYPE_CONTENT_INSTREAM_POSITION.

ContentInstreamPositionTargetingOptionDetails
Represents a targetable content instream position, which could be used by video and audio ads. This will be populated in the content_instream_position_details field when targeting_type is TARGETING_TYPE_CONTENT_INSTREAM_POSITION.

ContentOutstreamPositionAssignedTargetingOptionDetails
Assigned content outstream position targeting option details. This will be populated in the content_outstream_position_details field when targeting_type is TARGETING_TYPE_CONTENT_OUTSTREAM_POSITION.

ContentOutstreamPositionTargetingOptionDetails
Represents a targetable content outstream position, which could be used by display and video ads. This will be populated in the content_outstream_position_details field when targeting_type is TARGETING_TYPE_CONTENT_OUTSTREAM_POSITION.

ConversionCountingConfig
Settings that control how conversions are counted. All post-click conversions will be counted. A percentage value can be set for post-view conversions counting.

CounterEvent
Counter event of the creative.

CreateAssetRequest
A request message for CreateAsset.

CreateAssetResponse
A response message for CreateAsset.

CreateAssignedTargetingOptionsRequest
A request listing which assigned targeting options of a given targeting type should be created and added.

CreateSdfDownloadTaskRequest
Request message for [SdfDownloadTaskService.CreateSdfDownloadTask].

Creative
A single Creative.

CreativeConfig
Creative requirements configuration for the inventory source.

CustomBiddingAlgorithm
A single custom bidding algorithm.

CustomBiddingAlgorithmGetCall
Gets a custom bidding algorithm.

CustomBiddingAlgorithmListCall
Lists custom bidding algorithms that are accessible to the current user and can be used in bidding stratgies. The order is defined by the order_by parameter.

CustomBiddingAlgorithmMethods
A builder providing access to all methods supported on customBiddingAlgorithm resources. It is not used directly, but through the DisplayVideo hub.

CustomList
Describes a custom list entity, such as a custom affinity or custom intent audience list.

CustomListGetCall
Gets a custom list.

CustomListGroup
Details of custom list group. All custom list targeting settings are logically ‘OR’ of each other.

CustomListListCall
Lists custom lists. The order is defined by the order_by parameter.

CustomListMethods
A builder providing access to all methods supported on customList resources. It is not used directly, but through the DisplayVideo hub.

CustomListTargetingSetting
Details of custom list targeting setting.

Date
Represents a whole or partial calendar date, such as a birthday. The time of day and time zone are either specified elsewhere or are insignificant. The date is relative to the Gregorian Calendar. This can represent one of the following: * A full date, with non-zero year, month, and day values * A month and day value, with a zero year, such as an anniversary * A year on its own, with zero month and day values * A year and month value, with a zero day, such as a credit card expiration date Related types are google.type.TimeOfDay and google.protobuf.Timestamp.

DateRange
A date range.

DayAndTimeAssignedTargetingOptionDetails
Representation of a segment of time defined on a specific day of the week and with a start and end time. The time represented by start_hour must be before the time represented by end_hour.

DeactivateManualTriggerRequest
Request message for ManualTriggerService.DeactivateManualTrigger.

DeleteAssignedTargetingOptionsRequest
A request listing which assigned targeting options of a given targeting type should be deleted.

DeviceMakeModelAssignedTargetingOptionDetails
Assigned device make and model targeting option details. This will be populated in the device_make_model_details field when targeting_type is TARGETING_TYPE_DEVICE_MAKE_MODEL.

DeviceMakeModelTargetingOptionDetails
Represents a targetable device make and model. This will be populated in the device_make_model_details field of a TargetingOption when targeting_type is TARGETING_TYPE_DEVICE_MAKE_MODEL.

DeviceTypeAssignedTargetingOptionDetails
Targeting details for device type. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_DEVICE_TYPE.

DeviceTypeTargetingOptionDetails
Represents a targetable device type. This will be populated in the device_type_details field of a TargetingOption when targeting_type is TARGETING_TYPE_DEVICE_TYPE.

DigitalContentLabelAssignedTargetingOptionDetails
Targeting details for digital content label. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_DIGITAL_CONTENT_LABEL_EXCLUSION.

DigitalContentLabelTargetingOptionDetails
Represents a targetable digital content label rating tier. This will be populated in the digital_content_label_details field of the TargetingOption when targeting_type is TARGETING_TYPE_DIGITAL_CONTENT_LABEL_EXCLUSION.

Dimensions
Dimensions.

DisplayVideo
Central instance to access all DisplayVideo related resource activities

DoubleVerify
Details of DoubleVerify settings.

DoubleVerifyAppStarRating
Details of DoubleVerify star ratings settings.

DoubleVerifyBrandSafetyCategories
Settings for brand safety controls.

DoubleVerifyDisplayViewability
Details of DoubleVerify display viewability settings.

DoubleVerifyFraudInvalidTraffic
DoubleVerify Fraud & Invalid Traffic settings.

DoubleVerifyVideoViewability
Details of DoubleVerify video viewability settings.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

EnvironmentAssignedTargetingOptionDetails
Assigned environment targeting option details. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_ENVIRONMENT.

EnvironmentTargetingOptionDetails
Represents a targetable environment. This will be populated in the environment_details field of a TargetingOption when targeting_type is TARGETING_TYPE_ENVIRONMENT.

ExchangeAssignedTargetingOptionDetails
Details for assigned exchange targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_EXCHANGE.

ExchangeConfig
Settings that control which exchanges are enabled for a partner.

ExchangeConfigEnabledExchange
An enabled exchange in the partner.

ExchangeReviewStatus
Exchange review status for the creative.

ExchangeTargetingOptionDetails
Represents a targetable exchange. This will be populated in the exchange_details field of a TargetingOption when targeting_type is TARGETING_TYPE_EXCHANGE.

ExitEvent
Exit event of the creative.

FirstAndThirdPartyAudience
Describes a first or third party audience list used for targeting. First party audiences are created via usage of client data. Third party audiences are provided by Third Party data providers and can only be licensed to customers.

FirstAndThirdPartyAudienceGetCall
Gets a first and third party audience.

FirstAndThirdPartyAudienceGroup
Details of first and third party audience group. All first and third party audience targeting settings are logically ‘OR’ of each other.

FirstAndThirdPartyAudienceListCall
Lists first and third party audiences. The order is defined by the order_by parameter.

FirstAndThirdPartyAudienceMethods
A builder providing access to all methods supported on firstAndThirdPartyAudience resources. It is not used directly, but through the DisplayVideo hub.

FirstAndThirdPartyAudienceTargetingSetting
Details of first and third party audience targeting setting.

FixedBidStrategy
A strategy that uses a fixed bidding price.

FloodlightGroup
A single Floodlight group.

FloodlightGroupGetCall
Gets a Floodlight group.

FloodlightGroupMethods
A builder providing access to all methods supported on floodlightGroup resources. It is not used directly, but through the DisplayVideo hub.

FloodlightGroupPatchCall
Updates an existing Floodlight group. Returns the updated Floodlight group if successful.

FrequencyCap
Settings that control the number of times a user may be shown with the same ad during a given time period.

GenderAssignedTargetingOptionDetails
Details for assigned gender targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARTGETING_TYPE_GENDER.

GenderTargetingOptionDetails
Represents a targetable gender. This will be populated in the gender_details field of a TargetingOption when targeting_type is TARGETING_TYPE_GENDER.

GenerateDefaultLineItemRequest
Request message for LineItemService.GenerateDefaultLineItem.

GeoRegionAssignedTargetingOptionDetails
Details for assigned geographic region targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_GEO_REGION.

GeoRegionSearchTerms
Search terms for geo region targeting options.

GeoRegionTargetingOptionDetails
Represents a targetable geographic region. This will be populated in the geo_region_details field when targeting_type is TARGETING_TYPE_GEO_REGION.

GoogleAudience
Describes a Google audience resource. Includes Google audience lists.

GoogleAudienceGetCall
Gets a Google audience.

GoogleAudienceGroup
Details of Google audience group. All Google audience targeting settings are logically ‘OR’ of each other.

GoogleAudienceListCall
Lists Google audiences. The order is defined by the order_by parameter.

GoogleAudienceMethods
A builder providing access to all methods supported on googleAudience resources. It is not used directly, but through the DisplayVideo hub.

GoogleAudienceTargetingSetting
Details of Google audience targeting setting.

GoogleBytestreamMedia
Media resource.

HouseholdIncomeAssignedTargetingOptionDetails
Details for assigned household income targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_HOUSEHOLD_INCOME.

HouseholdIncomeTargetingOptionDetails
Represents a targetable household income. This will be populated in the household_income_details field of a TargetingOption when targeting_type is TARGETING_TYPE_HOUSEHOLD_INCOME.

IdFilter
A filtering option that filters entities by their entity IDs.

InsertionOrder
A single insertion order.

InsertionOrderBudget
Settings that control how insertion order budget is allocated.

InsertionOrderBudgetSegment
Settings that control the budget of a single budget segment.

IntegralAdScience
Details of Integral Ad Science settings.

IntegrationDetails
Integration details of an entry.

InventorySource
An inventory source.

InventorySourceAssignedTargetingOptionDetails
Targeting details for inventory source. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_INVENTORY_SOURCE.

InventorySourceDisplayCreativeConfig
The configuration for display creatives.

InventorySourceFilter
A filtering option for filtering on Inventory Source entities.

InventorySourceGetCall
Gets an inventory source.

InventorySourceGroup
A collection of targetable inventory sources.

InventorySourceGroupAssignedInventorySourceBulkEditCall
Bulk edits multiple assignments between inventory sources and a single inventory source group. The operation will delete the assigned inventory sources provided in BulkEditAssignedInventorySourcesRequest.deleted_assigned_inventory_sources and then create the assigned inventory sources provided in BulkEditAssignedInventorySourcesRequest.created_assigned_inventory_sources.

InventorySourceGroupAssignedInventorySourceCreateCall
Creates an assignment between an inventory source and an inventory source group.

InventorySourceGroupAssignedInventorySourceDeleteCall
Deletes the assignment between an inventory source and an inventory source group.

InventorySourceGroupAssignedInventorySourceListCall
Lists inventory sources assigned to an inventory source group.

InventorySourceGroupAssignedTargetingOptionDetails
Targeting details for inventory source group. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_INVENTORY_SOURCE_GROUP.

InventorySourceGroupCreateCall
Creates a new inventory source group. Returns the newly created inventory source group if successful.

InventorySourceGroupDeleteCall
Deletes an inventory source group.

InventorySourceGroupGetCall
Gets an inventory source group.

InventorySourceGroupListCall
Lists inventory source groups that are accessible to the current user. The order is defined by the order_by parameter.

InventorySourceGroupMethods
A builder providing access to all methods supported on inventorySourceGroup resources. It is not used directly, but through the DisplayVideo hub.

InventorySourceGroupPatchCall
Updates an inventory source group. Returns the updated inventory source group if successful.

InventorySourceListCall
Lists inventory sources that are accessible to the current user. The order is defined by the order_by parameter. If a filter by entity_status is not specified, inventory sources with entity status ENTITY_STATUS_ARCHIVED will not be included in the results.

InventorySourceMethods
A builder providing access to all methods supported on inventorySource resources. It is not used directly, but through the DisplayVideo hub.

InventorySourceStatus
The status related settings of the inventory source.

InventorySourceVideoCreativeConfig
The configuration for video creatives.

KeywordAssignedTargetingOptionDetails
Details for assigned keyword targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_KEYWORD.

LanguageAssignedTargetingOptionDetails
Details for assigned language targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_LANGUAGE.

LanguageTargetingOptionDetails
Represents a targetable language. This will be populated in the language_details field when targeting_type is TARGETING_TYPE_LANGUAGE.

LineItem
A single line item.

LineItemBudget
Settings that control how budget is allocated.

LineItemFlight
Settings that control the active duration of a line item.

ListAdvertiserAssignedTargetingOptionsResponse
Response message for ListAdvertiserAssignedTargetingOptions.

ListAdvertisersResponse
There is no detailed description.

ListAssignedInventorySourcesResponse
Response message for AssignedInventorySourceService.ListAssignedInventorySources.

ListAssignedLocationsResponse
Response message for AssignedLocationService.ListAssignedLocations.

ListCampaignAssignedTargetingOptionsResponse
Response message for ListCampaignAssignedTargetingOptions.

ListCampaignsResponse
There is no detailed description.

ListChannelsResponse
There is no detailed description.

ListCombinedAudiencesResponse
There is no detailed description.

ListCreativesResponse
There is no detailed description.

ListCustomBiddingAlgorithmsResponse
There is no detailed description.

ListCustomListsResponse
There is no detailed description.

ListFirstAndThirdPartyAudiencesResponse
There is no detailed description.

ListGoogleAudiencesResponse
There is no detailed description.

ListInsertionOrderAssignedTargetingOptionsResponse
Response message for ListInsertionOrderAssignedTargetingOptions.

ListInsertionOrdersResponse
There is no detailed description.

ListInventorySourceGroupsResponse
Response message for InventorySourceGroupService.ListInventorySourceGroups.

ListInventorySourcesResponse
There is no detailed description.

ListLineItemAssignedTargetingOptionsResponse
Response message for ListLineItemAssignedTargetingOptions.

ListLineItemsResponse
There is no detailed description.

ListLocationListsResponse
There is no detailed description.

ListManualTriggersResponse
There is no detailed description.

ListNegativeKeywordListsResponse
Response message for NegativeKeywordListService.ListNegativeKeywordLists.

ListNegativeKeywordsResponse
Response message for NegativeKeywordService.ListNegativeKeywords.

ListPartnerAssignedTargetingOptionsResponse
There is no detailed description.

ListPartnersResponse
There is no detailed description.

ListSitesResponse
Response message for SiteService.ListSites.

ListTargetingOptionsResponse
Response message for ListTargetingOptions.

ListUsersResponse
There is no detailed description.

LocationList
A list of locations used for targeting.

LookbackWindow
Specifies how many days into the past to look when determining whether to record a conversion.

ManualTrigger
A single manual trigger in Display & Video 360.

MaximizeSpendBidStrategy
A strategy that automatically adjusts the bid to optimize a specified performance goal while spending the full budget.

MeasurementConfig
Measurement settings of a partner.

MediaDownloadCall
Downloads media. Download is supported on the URI /download/{resource_name=**}?alt=media. Note: Download requests will not be successful without including alt=media query string.

MediaMethods
A builder providing access to all methods supported on media resources. It is not used directly, but through the DisplayVideo hub.

MobileApp
A mobile app promoted by a mobile app install line item.

Money
Represents an amount of money with its currency type.

NegativeKeyword
A negatively targeted keyword that belongs to a negative keyword list.

NegativeKeywordList
A list of negative keywords used for targeting.

NegativeKeywordListAssignedTargetingOptionDetails
Targeting details for negative keyword list. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_NEGATIVE_KEYWORD_LIST.

ObaIcon
OBA Icon for a Creative

OnScreenPositionAssignedTargetingOptionDetails
On screen position targeting option details. This will be populated in the on_screen_position_details field when targeting_type is TARGETING_TYPE_ON_SCREEN_POSITION.

OnScreenPositionTargetingOptionDetails
Represents a targetable on screen position, which could be used by display and video ads. This will be populated in the on_screen_position_details field when targeting_type is TARGETING_TYPE_ON_SCREEN_POSITION.

OperatingSystemAssignedTargetingOptionDetails
Assigned operating system targeting option details. This will be populated in the operating_system_details field when targeting_type is TARGETING_TYPE_OPERATING_SYSTEM.

OperatingSystemTargetingOptionDetails
Represents a targetable operating system. This will be populated in the operating_system_details field of a TargetingOption when targeting_type is TARGETING_TYPE_OPERATING_SYSTEM.

Operation
This resource represents a long-running operation that is the result of a network API call.

Pacing
Settings that control the rate at which a budget is spent.

ParentEntityFilter
A filtering option that filters on selected file types belonging to a chosen set of filter entities.

ParentalStatusAssignedTargetingOptionDetails
Details for assigned parental status targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARTGETING_TYPE_PARENTAL_STATUS.

ParentalStatusTargetingOptionDetails
Represents a targetable parental status. This will be populated in the parental_status_details field of a TargetingOption when targeting_type is TARGETING_TYPE_PARENTAL_STATUS.

Partner
A single partner in Display & Video 360 (DV360).

PartnerAdServerConfig
Ad server related settings of a partner.

PartnerBulkEditPartnerAssignedTargetingOptionCall
Bulk edits targeting options under a single partner. The operation will delete the assigned targeting options provided in BulkEditPartnerAssignedTargetingOptionsRequest.deleteRequests and then create the assigned targeting options provided in BulkEditPartnerAssignedTargetingOptionsRequest.createRequests .

PartnerChannelCreateCall
Creates a new channel. Returns the newly created channel if successful.

PartnerChannelGetCall
Gets a channel for a partner or advertiser.

PartnerChannelListCall
Lists channels for a partner or advertiser.

PartnerChannelPatchCall
Updates a channel. Returns the updated channel if successful.

PartnerChannelSiteBulkEditCall
Bulk edits sites under a single channel. The operation will delete the sites provided in BulkEditSitesRequest.deleted_sites and then create the sites provided in BulkEditSitesRequest.created_sites.

PartnerChannelSiteCreateCall
Creates a site in a channel.

PartnerChannelSiteDeleteCall
Deletes a site from a channel.

PartnerChannelSiteListCall
Lists sites in a channel.

PartnerChannelSiteReplaceCall
Replaces all of the sites under a single channel. The operation will replace the sites under a channel with the sites provided in ReplaceSitesRequest.new_sites.

PartnerCost
Settings that control a partner cost. A partner cost is any type of expense involved in running a campaign, other than the costs of purchasing impressions (which is called the media cost) and using third-party audience segment data (data fee). Some examples of partner costs include the fees for using DV360, a third-party ad server, or a third-party ad serving verification service.

PartnerDataAccessConfig
Settings that control how partner related data may be accessed.

PartnerGeneralConfig
General settings of a partner.

PartnerGetCall
Gets a partner.

PartnerListCall
Lists partners that are accessible to the current user. The order is defined by the order_by parameter.

PartnerMethods
A builder providing access to all methods supported on partner resources. It is not used directly, but through the DisplayVideo hub.

PartnerRevenueModel
Settings that control how partner revenue is calculated.

PartnerTargetingTypeAssignedTargetingOptionCreateCall
Assigns a targeting option to a partner. Returns the assigned targeting option if successful.

PartnerTargetingTypeAssignedTargetingOptionDeleteCall
Deletes an assigned targeting option from a partner.

PartnerTargetingTypeAssignedTargetingOptionGetCall
Gets a single targeting option assigned to a partner.

PartnerTargetingTypeAssignedTargetingOptionListCall
Lists the targeting options assigned to a partner.

PerformanceGoal
Settings that control the performance goal of a campaign or insertion order.

PerformanceGoalBidStrategy
A strategy that automatically adjusts the bid to meet or beat a specified performance goal.

ProximityLocationListAssignedTargetingOptionDetails
Targeting details for proximity location list. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_PROXIMITY_LOCATION_LIST.

PublisherReviewStatus
Publisher review status for the creative.

RateDetails
The rate related settings of the inventory source.

RegionalLocationListAssignedTargetingOptionDetails
Targeting details for regional location list. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_REGIONAL_LOCATION_LIST.

ReplaceNegativeKeywordsRequest
Request message for NegativeKeywordService.ReplaceNegativeKeywords.

ReplaceNegativeKeywordsResponse
Response message for NegativeKeywordService.ReplaceNegativeKeywords.

ReplaceSitesRequest
Request message for SiteService.ReplaceSites.

ReplaceSitesResponse
Response message for SiteService.ReplaceSites.

ReviewStatusInfo
Review statuses for the creative.

SdfConfig
Structured Data File (SDF) related settings.

SdfdownloadtaskCreateCall
Creates an SDF Download Task. Returns an Operation. An SDF Download Task is a long-running, asynchronous operation. The metadata type of this operation is SdfDownloadTaskMetadata. If the request is successful, the response type of the operation is SdfDownloadTask. The response will not include the download files, which must be retrieved with media.download. The state of operation can be retrieved with sdfdownloadtask.operations.get. Any errors can be found in the error.message. Note that error.details is expected to be empty.

SdfdownloadtaskMethods
A builder providing access to all methods supported on sdfdownloadtask resources. It is not used directly, but through the DisplayVideo hub.

SdfdownloadtaskOperationGetCall
Gets the latest state of an asynchronous SDF download task operation. Clients should poll this method at intervals of 30 seconds.

SearchTargetingOptionsRequest
Request message for SearchTargetingOptions.

SearchTargetingOptionsResponse
Response message for SearchTargetingOptionsResponse.

SensitiveCategoryAssignedTargetingOptionDetails
Targeting details for sensitive category. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_SENSITIVE_CATEGORY_EXCLUSION.

SensitiveCategoryTargetingOptionDetails
Represents a targetable sensitive category. This will be populated in the sensitive_category_details field of the TargetingOption when targeting_type is TARGETING_TYPE_SENSITIVE_CATEGORY_EXCLUSION.

Site
A single site. Sites are apps or websites belonging to a channel.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

SubExchangeAssignedTargetingOptionDetails
Details for assigned sub-exchange targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_SUB_EXCHANGE.

SubExchangeTargetingOptionDetails
Represents a targetable sub-exchange. This will be populated in the sub_exchange_details field of a TargetingOption when targeting_type is TARGETING_TYPE_SUB_EXCHANGE.

TargetingExpansionConfig
Settings that control the targeting expansion of the line item. Targeting expansion allows the line item to reach a larger audience based on the original audience list and the targeting expansion level.

TargetingOption
Represents a single targeting option, which is a targetable concept in DV360.

TargetingTypeMethods
A builder providing access to all methods supported on targetingType resources. It is not used directly, but through the DisplayVideo hub.

TargetingTypeTargetingOptionGetCall
Gets a single targeting option.

TargetingTypeTargetingOptionListCall
Lists targeting options of a given type.

TargetingTypeTargetingOptionSearchCall
Searches for targeting options of a given type based on the given search terms.

ThirdPartyOnlyConfig
Settings for advertisers that use third-party ad servers only.

ThirdPartyUrl
Tracking URLs from third parties to track interactions with an audio or a video creative.

ThirdPartyVerifierAssignedTargetingOptionDetails
Assigned third party verifier targeting option details. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_THIRD_PARTY_VERIFIER.

TimeRange
A time range.

TimerEvent
Timer event of the creative.

TrackingFloodlightActivityConfig
Settings that control the behavior of a single Floodlight activity config.

Transcode
Represents information about the transcoded audio or video file.

UniversalAdId
A creative identifier provided by a registry that is unique across all platforms. This is part of the VAST 4.0 standard.

UrlAssignedTargetingOptionDetails
Details for assigned URL targeting option. This will be populated in the details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_URL.

User
A single user in Display & Video 360.

UserBulkEditAssignedUserRoleCall
Bulk edits user roles for a user. The operation will delete the assigned user roles provided in BulkEditAssignedUserRolesRequest.deletedAssignedUserRoles and then assign the user roles provided in BulkEditAssignedUserRolesRequest.createdAssignedUserRoles.

UserCreateCall
Creates a new user. Returns the newly created user if successful.

UserDeleteCall
Deletes a user.

UserGetCall
Gets a user.

UserListCall
Lists users that are accessible to the current user. If two users have user roles on the same partner or advertiser, they can access each other.

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the DisplayVideo hub.

UserPatchCall
Updates an existing user. Returns the updated user if successful.

UserRewardedContentAssignedTargetingOptionDetails
User rewarded content targeting option details. This will be populated in the user_rewarded_content_details field when targeting_type is TARGETING_TYPE_USER_REWARDED_CONTENT.

UserRewardedContentTargetingOptionDetails
Represents a targetable user rewarded content status for video ads only. This will be populated in the user_rewarded_content_details field when targeting_type is TARGETING_TYPE_USER_REWARDED_CONTENT.

VideoPlayerSizeAssignedTargetingOptionDetails
Video player size targeting option details. This will be populated in the video_player_size_details field when targeting_type is TARGETING_TYPE_VIDEO_PLAYER_SIZE. Explicitly targeting all options is not supported. Remove all video player size targeting options to achieve this effect.

VideoPlayerSizeTargetingOptionDetails
Represents a targetable video player size. This will be populated in the video_player_size_details field when targeting_type is TARGETING_TYPE_VIDEO_PLAYER_SIZE.

ViewabilityAssignedTargetingOptionDetails
Assigned viewability targeting option details. This will be populated in the viewability_details field of an AssignedTargetingOption when targeting_type is TARGETING_TYPE_VIEWABILITY.

ViewabilityTargetingOptionDetails
Represents a targetable viewability. This will be populated in the viewability_details field of a TargetingOption when targeting_type is TARGETING_TYPE_VIEWABILITY.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.