Struct google_displayvideo1::api::Advertiser

pub struct Advertiser {
    pub ad_server_config: Option<AdvertiserAdServerConfig>,
    pub advertiser_id: Option<String>,
    pub creative_config: Option<AdvertiserCreativeConfig>,
    pub data_access_config: Option<AdvertiserDataAccessConfig>,
    pub display_name: Option<String>,
    pub entity_status: Option<String>,
    pub general_config: Option<AdvertiserGeneralConfig>,
    pub integration_details: Option<IntegrationDetails>,
    pub name: Option<String>,
    pub partner_id: Option<String>,
    pub serving_config: Option<AdvertiserTargetingConfig>,
    pub update_time: Option<String>,
}

A single advertiser in Display & Video 360 (DV360).

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

assets upload advertisers (none)

campaigns targeting types assigned targeting options get advertisers (none)

campaigns targeting types assigned targeting options list advertisers (none)

campaigns bulk list campaign assigned targeting options advertisers (none)

campaigns create advertisers (none)

campaigns delete advertisers (none)

campaigns get advertisers (none)

campaigns list advertisers (none)

campaigns patch advertisers (none)

channels sites bulk edit advertisers (none)

channels sites create advertisers (none)

channels sites delete advertisers (none)

channels sites list advertisers (none)

channels sites replace advertisers (none)

channels create advertisers (none)

channels get advertisers (none)

channels list advertisers (none)

channels patch advertisers (none)

creatives create advertisers (none)

creatives delete advertisers (none)

creatives get advertisers (none)

creatives list advertisers (none)

creatives patch advertisers (none)

insertion orders targeting types assigned targeting options get advertisers (none)

insertion orders targeting types assigned targeting options list advertisers (none)

insertion orders bulk list insertion order assigned targeting options advertisers (none)

insertion orders create advertisers (none)

insertion orders delete advertisers (none)

insertion orders get advertisers (none)

insertion orders list advertisers (none)

insertion orders patch advertisers (none)

line items targeting types assigned targeting options create advertisers (none)

line items targeting types assigned targeting options delete advertisers (none)

line items targeting types assigned targeting options get advertisers (none)

line items targeting types assigned targeting options list advertisers (none)

line items bulk edit line item assigned targeting options advertisers (none)

line items bulk list line item assigned targeting options advertisers (none)

line items create advertisers (none)

line items delete advertisers (none)

line items generate default advertisers (none)

line items get advertisers (none)

line items list advertisers (none)

line items patch advertisers (none)

location lists assigned locations bulk edit advertisers (none)

location lists assigned locations create advertisers (none)

location lists assigned locations delete advertisers (none)

location lists assigned locations list advertisers (none)

location lists create advertisers (none)

location lists get advertisers (none)

location lists list advertisers (none)

location lists patch advertisers (none)

manual triggers activate advertisers (none)

manual triggers create advertisers (none)

manual triggers deactivate advertisers (none)

manual triggers get advertisers (none)

manual triggers list advertisers (none)

manual triggers patch advertisers (none)

negative keyword lists negative keywords bulk edit advertisers (none)

negative keyword lists negative keywords create advertisers (none)

negative keyword lists negative keywords delete advertisers (none)

negative keyword lists negative keywords list advertisers (none)

negative keyword lists negative keywords replace advertisers (none)

negative keyword lists create advertisers (none)

negative keyword lists delete advertisers (none)

negative keyword lists get advertisers (none)

negative keyword lists list advertisers (none)

negative keyword lists patch advertisers (none)

targeting types assigned targeting options create advertisers (none)

targeting types assigned targeting options delete advertisers (none)

targeting types assigned targeting options get advertisers (none)

targeting types assigned targeting options list advertisers (none)

audit advertisers (none)

bulk edit advertiser assigned targeting options advertisers (none)

bulk list advertiser assigned targeting options advertisers (none)

create advertisers (request|response)

delete advertisers (none)

get advertisers (response)

list advertisers (none)

patch advertisers (request|response)

Fields
ad_server_config: Option<AdvertiserAdServerConfig>

Required. Immutable. Ad server related settings of the advertiser.

advertiser_id: Option<String>

Output only. The unique ID of the advertiser. Assigned by the system.

creative_config: Option<AdvertiserCreativeConfig>

Required. Creative related settings of the advertiser.

data_access_config: Option<AdvertiserDataAccessConfig>

Settings that control how advertiser data may be accessed.

display_name: Option<String>

Required. The display name of the advertiser. Must be UTF-8 encoded with a maximum size of 240 bytes.

entity_status: Option<String>

Required. Controls whether or not insertion orders and line items of the advertiser can spend their budgets and bid on inventory. * Accepted values are ENTITY_STATUS_ACTIVE and ENTITY_STATUS_SCHEDULED_FOR_DELETION. * If set to ENTITY_STATUS_SCHEDULED_FOR_DELETION, the advertiser will be deleted 30 days from when it was first scheduled for deletion.

general_config: Option<AdvertiserGeneralConfig>

Required. General settings of the advertiser.

integration_details: Option<IntegrationDetails>

Integration details of the advertiser. Only integrationCode is currently applicable to advertiser. Other fields of IntegrationDetails are not supported and will be ignored if provided.

name: Option<String>

Output only. The resource name of the advertiser.

partner_id: Option<String>

Required. Immutable. The unique ID of the partner that the advertiser belongs to.

serving_config: Option<AdvertiserTargetingConfig>

Targeting settings related to ad serving of the advertiser.

update_time: Option<String>

Output only. The timestamp when the advertiser was last updated. Assigned by the system.