Struct google_displayvideo1::api::AdvertiserMethods

pub struct AdvertiserMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on advertiser resources. It is not used directly, but through the DisplayVideo hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_displayvideo1 as displayvideo1;
 
use std::default::Default;
use oauth2;
use displayvideo1::DisplayVideo;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = DisplayVideo::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `assets_upload(...)`, `audit(...)`, `bulk_edit_advertiser_assigned_targeting_options(...)`, `bulk_list_advertiser_assigned_targeting_options(...)`, `campaigns_bulk_list_campaign_assigned_targeting_options(...)`, `campaigns_create(...)`, `campaigns_delete(...)`, `campaigns_get(...)`, `campaigns_list(...)`, `campaigns_patch(...)`, `campaigns_targeting_types_assigned_targeting_options_get(...)`, `campaigns_targeting_types_assigned_targeting_options_list(...)`, `channels_create(...)`, `channels_get(...)`, `channels_list(...)`, `channels_patch(...)`, `channels_sites_bulk_edit(...)`, `channels_sites_create(...)`, `channels_sites_delete(...)`, `channels_sites_list(...)`, `channels_sites_replace(...)`, `create(...)`, `creatives_create(...)`, `creatives_delete(...)`, `creatives_get(...)`, `creatives_list(...)`, `creatives_patch(...)`, `delete(...)`, `get(...)`, `insertion_orders_bulk_list_insertion_order_assigned_targeting_options(...)`, `insertion_orders_create(...)`, `insertion_orders_delete(...)`, `insertion_orders_get(...)`, `insertion_orders_list(...)`, `insertion_orders_patch(...)`, `insertion_orders_targeting_types_assigned_targeting_options_get(...)`, `insertion_orders_targeting_types_assigned_targeting_options_list(...)`, `line_items_bulk_edit_line_item_assigned_targeting_options(...)`, `line_items_bulk_list_line_item_assigned_targeting_options(...)`, `line_items_create(...)`, `line_items_delete(...)`, `line_items_generate_default(...)`, `line_items_get(...)`, `line_items_list(...)`, `line_items_patch(...)`, `line_items_targeting_types_assigned_targeting_options_create(...)`, `line_items_targeting_types_assigned_targeting_options_delete(...)`, `line_items_targeting_types_assigned_targeting_options_get(...)`, `line_items_targeting_types_assigned_targeting_options_list(...)`, `list(...)`, `location_lists_assigned_locations_bulk_edit(...)`, `location_lists_assigned_locations_create(...)`, `location_lists_assigned_locations_delete(...)`, `location_lists_assigned_locations_list(...)`, `location_lists_create(...)`, `location_lists_get(...)`, `location_lists_list(...)`, `location_lists_patch(...)`, `manual_triggers_activate(...)`, `manual_triggers_create(...)`, `manual_triggers_deactivate(...)`, `manual_triggers_get(...)`, `manual_triggers_list(...)`, `manual_triggers_patch(...)`, `negative_keyword_lists_create(...)`, `negative_keyword_lists_delete(...)`, `negative_keyword_lists_get(...)`, `negative_keyword_lists_list(...)`, `negative_keyword_lists_negative_keywords_bulk_edit(...)`, `negative_keyword_lists_negative_keywords_create(...)`, `negative_keyword_lists_negative_keywords_delete(...)`, `negative_keyword_lists_negative_keywords_list(...)`, `negative_keyword_lists_negative_keywords_replace(...)`, `negative_keyword_lists_patch(...)`, `patch(...)`, `targeting_types_assigned_targeting_options_create(...)`, `targeting_types_assigned_targeting_options_delete(...)`, `targeting_types_assigned_targeting_options_get(...)` and `targeting_types_assigned_targeting_options_list(...)`
// to build up your call.
let rb = hub.advertisers();

Implementations

impl<'a, C> AdvertiserMethods<'a, C>

pub fn assets_upload(
&self,
request: CreateAssetRequest,
advertiser_id: &str
) -> AdvertiserAssetUploadCall<'a, C>

Create a builder to help you perform the following task:

Uploads an asset. Returns the ID of the newly uploaded asset if successful. The asset file size should be no more than 10 MB for images, 200 MB for ZIP files, and 1 GB for videos.

Arguments

request - No description provided.

advertiserId - Required. The ID of the advertiser this asset belongs to.

pub fn campaigns_targeting_types_assigned_targeting_options_get(
&self,
advertiser_id: &str,
campaign_id: &str,
targeting_type: &str,
assigned_targeting_option_id: &str
) -> AdvertiserCampaignTargetingTypeAssignedTargetingOptionGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a single targeting option assigned to a campaign.

Arguments

advertiserId - Required. The ID of the advertiser the campaign belongs to.

campaignId - Required. The ID of the campaign the assigned targeting option belongs to.

targetingType - Required. Identifies the type of this assigned targeting option. Supported targeting types: * TARGETING_TYPE_AGE_RANGE * TARGETING_TYPE_AUTHORIZED_SELLER_STATUS * TARGETING_TYPE_CONTENT_INSTREAM_POSITION * TARGETING_TYPE_CONTENT_OUTSTREAM_POSITION * TARGETING_TYPE_DIGITAL_CONTENT_LABEL_EXCLUSION * TARGETING_TYPE_ENVIRONMENT * TARGETING_TYPE_EXCHANGE * TARGETING_TYPE_GENDER * TARGETING_TYPE_GEO_REGION * TARGETING_TYPE_HOUSEHOLD_INCOME * TARGETING_TYPE_INVENTORY_SOURCE * TARGETING_TYPE_INVENTORY_SOURCE_GROUP * TARGETING_TYPE_LANGUAGE * TARGETING_TYPE_ON_SCREEN_POSITION * TARGETING_TYPE_PARENTAL_STATUS * TARGETING_TYPE_SENSITIVE_CATEGORY_EXCLUSION * TARGETING_TYPE_SUB_EXCHANGE * TARGETING_TYPE_THIRD_PARTY_VERIFIER * TARGETING_TYPE_VIEWABILITY

assignedTargetingOptionId - Required. An identifier unique to the targeting type in this campaign that identifies the assigned targeting option being requested.

pub fn campaigns_targeting_types_assigned_targeting_options_list(
&self,
advertiser_id: &str,
campaign_id: &str,
targeting_type: &str
) -> AdvertiserCampaignTargetingTypeAssignedTargetingOptionListCall<'a, C>

Create a builder to help you perform the following task:

Lists the targeting options assigned to a campaign for a specified targeting type.

Arguments

advertiserId - Required. The ID of the advertiser the campaign belongs to.

campaignId - Required. The ID of the campaign to list assigned targeting options for.

targetingType - Required. Identifies the type of assigned targeting options to list. Supported targeting types: * TARGETING_TYPE_AGE_RANGE * TARGETING_TYPE_AUTHORIZED_SELLER_STATUS * TARGETING_TYPE_CONTENT_INSTREAM_POSITION * TARGETING_TYPE_CONTENT_OUTSTREAM_POSITION * TARGETING_TYPE_DIGITAL_CONTENT_LABEL_EXCLUSION * TARGETING_TYPE_ENVIRONMENT * TARGETING_TYPE_EXCHANGE * TARGETING_TYPE_GENDER * TARGETING_TYPE_GEO_REGION * TARGETING_TYPE_HOUSEHOLD_INCOME * TARGETING_TYPE_INVENTORY_SOURCE * TARGETING_TYPE_INVENTORY_SOURCE_GROUP * TARGETING_TYPE_LANGUAGE * TARGETING_TYPE_ON_SCREEN_POSITION * TARGETING_TYPE_PARENTAL_STATUS * TARGETING_TYPE_SENSITIVE_CATEGORY_EXCLUSION * TARGETING_TYPE_SUB_EXCHANGE * TARGETING_TYPE_THIRD_PARTY_VERIFIER * TARGETING_TYPE_VIEWABILITY

pub fn campaigns_bulk_list_campaign_assigned_targeting_options(
&self,
advertiser_id: &str,
campaign_id: &str
) -> AdvertiserCampaignBulkListCampaignAssignedTargetingOptionCall<'a, C>

Create a builder to help you perform the following task:

Lists assigned targeting options of a campaign across targeting types.

Arguments

advertiserId - Required. The ID of the advertiser the campaign belongs to.

campaignId - Required. The ID of the campaign to list assigned targeting options for.

pub fn campaigns_create(
&self,
request: Campaign,
advertiser_id: &str
) -> AdvertiserCampaignCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new campaign. Returns the newly created campaign if successful.

Arguments

request - No description provided.

advertiserId - Output only. The unique ID of the advertiser the campaign belongs to.

pub fn campaigns_delete(
&self,
advertiser_id: &str,
campaign_id: &str
) -> AdvertiserCampaignDeleteCall<'a, C>

Create a builder to help you perform the following task:

Permanently deletes a campaign. A deleted campaign cannot be recovered. The campaign should be archived first, i.e. set entity_status to ENTITY_STATUS_ARCHIVED, to be able to delete it.

Arguments

advertiserId - The ID of the advertiser this campaign belongs to.

campaignId - The ID of the campaign we need to delete.

pub fn campaigns_get(
&self,
advertiser_id: &str,
campaign_id: &str
) -> AdvertiserCampaignGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a campaign.

Arguments

advertiserId - Required. The ID of the advertiser this campaign belongs to.

campaignId - Required. The ID of the campaign to fetch.

pub fn campaigns_list(
&self,
advertiser_id: &str
) -> AdvertiserCampaignListCall<'a, C>

Create a builder to help you perform the following task:

Lists campaigns in an advertiser. The order is defined by the order_by parameter. If a filter by entity_status is not specified, campaigns with ENTITY_STATUS_ARCHIVED will not be included in the results.

Arguments

advertiserId - The ID of the advertiser to list campaigns for.

pub fn campaigns_patch(
&self,
request: Campaign,
advertiser_id: &str,
campaign_id: &str
) -> AdvertiserCampaignPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing campaign. Returns the updated campaign if successful.

Arguments

request - No description provided.

advertiserId - Output only. The unique ID of the advertiser the campaign belongs to.

campaignId - Output only. The unique ID of the campaign. Assigned by the system.

pub fn channels_sites_bulk_edit(
&self,
request: BulkEditSitesRequest,
advertiser_id: &str,
channel_id: &str
) -> AdvertiserChannelSiteBulkEditCall<'a, C>

Create a builder to help you perform the following task:

Bulk edits sites under a single channel. The operation will delete the sites provided in BulkEditSitesRequest.deleted_sites and then create the sites provided in BulkEditSitesRequest.created_sites.

Arguments

request - No description provided.

advertiserId - The ID of the advertiser that owns the parent channel.

channelId - Required. The ID of the parent channel to which the sites belong.

pub fn channels_sites_create(
&self,
request: Site,
advertiser_id: &str,
channel_id: &str
) -> AdvertiserChannelSiteCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a site in a channel.

Arguments

request - No description provided.

advertiserId - The ID of the advertiser that owns the parent channel.

channelId - Required. The ID of the parent channel in which the site will be created.

pub fn channels_sites_delete(
&self,
advertiser_id: &str,
channel_id: &str,
url_or_app_id: &str
) -> AdvertiserChannelSiteDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a site from a channel.

Arguments

advertiserId - The ID of the advertiser that owns the parent channel.

channelId - Required. The ID of the parent channel to which the site belongs.

urlOrAppId - Required. The URL or app ID of the site to delete.

pub fn channels_sites_list(
&self,
advertiser_id: &str,
channel_id: &str
) -> AdvertiserChannelSiteListCall<'a, C>

Create a builder to help you perform the following task:

Lists sites in a channel.

Arguments

advertiserId - The ID of the advertiser that owns the parent channel.

channelId - Required. The ID of the parent channel to which the requested sites belong.

pub fn channels_sites_replace(
&self,
request: ReplaceSitesRequest,
advertiser_id: &str,
channel_id: &str
) -> AdvertiserChannelSiteReplaceCall<'a, C>

Create a builder to help you perform the following task:

Replaces all of the sites under a single channel. The operation will replace the sites under a channel with the sites provided in ReplaceSitesRequest.new_sites.

Arguments

request - No description provided.

advertiserId - The ID of the advertiser that owns the parent channel.

channelId - Required. The ID of the parent channel whose sites will be replaced.

pub fn channels_create(
&self,
request: Channel,
advertiser_id: &str
) -> AdvertiserChannelCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new channel. Returns the newly created channel if successful.

Arguments

request - No description provided.

advertiserId - The ID of the advertiser that owns the created channel.

pub fn channels_get(
&self,
advertiser_id: &str,
channel_id: &str
) -> AdvertiserChannelGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a channel for a partner or advertiser.

Arguments

advertiserId - The ID of the advertiser that owns the fetched channel.

channelId - Required. The ID of the channel to fetch.

pub fn channels_list(
&self,
advertiser_id: &str
) -> AdvertiserChannelListCall<'a, C>

Create a builder to help you perform the following task:

Lists channels for a partner or advertiser.

Arguments

advertiserId - The ID of the advertiser that owns the channels.

pub fn channels_patch(
&self,
request: Channel,
advertiser_id: &str,
channel_id: &str
) -> AdvertiserChannelPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a channel. Returns the updated channel if successful.

Arguments

request - No description provided.

advertiserId - The ID of the advertiser that owns the created channel.

channelId - Output only. The unique ID of the channel. Assigned by the system.

pub fn creatives_create(
&self,
request: Creative,
advertiser_id: &str
) -> AdvertiserCreativeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new creative. Returns the newly created creative if successful.

Arguments

request - No description provided.

advertiserId - Output only. The unique ID of the advertiser the creative belongs to.

pub fn creatives_delete(
&self,
advertiser_id: &str,
creative_id: &str
) -> AdvertiserCreativeDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a creative. Returns error code NOT_FOUND if the creative does not exist. The creative should be archived first, i.e. set entity_status to ENTITY_STATUS_ARCHIVED, before it can be deleted.

Arguments

advertiserId - The ID of the advertiser this creative belongs to.

creativeId - The ID of the creative to be deleted.

pub fn creatives_get(
&self,
advertiser_id: &str,
creative_id: &str
) -> AdvertiserCreativeGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a creative.

Arguments

advertiserId - Required. The ID of the advertiser this creative belongs to.

creativeId - Required. The ID of the creative to fetch.

pub fn creatives_list(
&self,
advertiser_id: &str
) -> AdvertiserCreativeListCall<'a, C>

Create a builder to help you perform the following task:

Lists creatives in an advertiser. The order is defined by the order_by parameter. If a filter by entity_status is not specified, creatives with ENTITY_STATUS_ARCHIVED will not be included in the results.

Arguments

advertiserId - Required. The ID of the advertiser to list creatives for.

pub fn creatives_patch(
&self,
request: Creative,
advertiser_id: &str,
creative_id: &str
) -> AdvertiserCreativePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing creative. Returns the updated creative if successful.

Arguments

request - No description provided.

advertiserId - Output only. The unique ID of the advertiser the creative belongs to.

creativeId - Output only. The unique ID of the creative. Assigned by the system.

pub fn insertion_orders_targeting_types_assigned_targeting_options_get(
&self,
advertiser_id: &str,
insertion_order_id: &str,
targeting_type: &str,
assigned_targeting_option_id: &str
) -> AdvertiserInsertionOrderTargetingTypeAssignedTargetingOptionGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a single targeting option assigned to an insertion order.

Arguments

advertiserId - Required. The ID of the advertiser the insertion order belongs to.

insertionOrderId - Required. The ID of the insertion order the assigned targeting option belongs to.

targetingType - Required. Identifies the type of this assigned targeting option.

assignedTargetingOptionId - Required. An identifier unique to the targeting type in this insertion order that identifies the assigned targeting option being requested.

pub fn insertion_orders_targeting_types_assigned_targeting_options_list(
&self,
advertiser_id: &str,
insertion_order_id: &str,
targeting_type: &str
) -> AdvertiserInsertionOrderTargetingTypeAssignedTargetingOptionListCall<'a, C>

Create a builder to help you perform the following task:

Lists the targeting options assigned to an insertion order.

Arguments

advertiserId - Required. The ID of the advertiser the insertion order belongs to.

insertionOrderId - Required. The ID of the insertion order to list assigned targeting options for.

targetingType - Required. Identifies the type of assigned targeting options to list.

pub fn insertion_orders_bulk_list_insertion_order_assigned_targeting_options(
&self,
advertiser_id: &str,
insertion_order_id: &str
) -> AdvertiserInsertionOrderBulkListInsertionOrderAssignedTargetingOptionCall<'a, C>

Create a builder to help you perform the following task:

Lists assigned targeting options of an insertion order across targeting types.

Arguments

advertiserId - Required. The ID of the advertiser the insertion order belongs to.

insertionOrderId - Required. The ID of the insertion order to list assigned targeting options for.

pub fn insertion_orders_create(
&self,
request: InsertionOrder,
advertiser_id: &str
) -> AdvertiserInsertionOrderCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new insertion order. Returns the newly created insertion order if successful.

Arguments

request - No description provided.

advertiserId - Output only. The unique ID of the advertiser the insertion order belongs to.

pub fn insertion_orders_delete(
&self,
advertiser_id: &str,
insertion_order_id: &str
) -> AdvertiserInsertionOrderDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an insertion order. Returns error code NOT_FOUND if the insertion order does not exist. The insertion order should be archived first, i.e. set entity_status to ENTITY_STATUS_ARCHIVED, to be able to delete it.

Arguments

advertiserId - The ID of the advertiser this insertion order belongs to.

insertionOrderId - The ID of the insertion order we need to delete.

pub fn insertion_orders_get(
&self,
advertiser_id: &str,
insertion_order_id: &str
) -> AdvertiserInsertionOrderGetCall<'a, C>

Create a builder to help you perform the following task:

Gets an insertion order. Returns error code NOT_FOUND if the insertion order does not exist.

Arguments

advertiserId - Required. The ID of the advertiser this insertion order belongs to.

insertionOrderId - Required. The ID of the insertion order to fetch.

pub fn insertion_orders_list(
&self,
advertiser_id: &str
) -> AdvertiserInsertionOrderListCall<'a, C>

Create a builder to help you perform the following task:

Lists insertion orders in an advertiser. The order is defined by the order_by parameter. If a filter by entity_status is not specified, insertion orders with ENTITY_STATUS_ARCHIVED will not be included in the results.

Arguments

advertiserId - Required. The ID of the advertiser to list insertion orders for.

pub fn insertion_orders_patch(
&self,
request: InsertionOrder,
advertiser_id: &str,
insertion_order_id: &str
) -> AdvertiserInsertionOrderPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing insertion order. Returns the updated insertion order if successful.

Arguments

request - No description provided.

advertiserId - Output only. The unique ID of the advertiser the insertion order belongs to.

insertionOrderId - Output only. The unique ID of the insertion order. Assigned by the system.

pub fn line_items_targeting_types_assigned_targeting_options_create(
&self,
request: AssignedTargetingOption,
advertiser_id: &str,
line_item_id: &str,
targeting_type: &str
) -> AdvertiserLineItemTargetingTypeAssignedTargetingOptionCreateCall<'a, C>

Create a builder to help you perform the following task:

Assigns a targeting option to a line item. Returns the assigned targeting option if successful.

Arguments

request - No description provided.

advertiserId - Required. The ID of the advertiser the line item belongs to.

lineItemId - Required. The ID of the line item the assigned targeting option will belong to.

targetingType - Required. Identifies the type of this assigned targeting option.

pub fn line_items_targeting_types_assigned_targeting_options_delete(
&self,
advertiser_id: &str,
line_item_id: &str,
targeting_type: &str,
assigned_targeting_option_id: &str
) -> AdvertiserLineItemTargetingTypeAssignedTargetingOptionDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an assigned targeting option from a line item.

Arguments

advertiserId - Required. The ID of the advertiser the line item belongs to.

lineItemId - Required. The ID of the line item the assigned targeting option belongs to.

targetingType - Required. Identifies the type of this assigned targeting option.

assignedTargetingOptionId - Required. The ID of the assigned targeting option to delete.

pub fn line_items_targeting_types_assigned_targeting_options_get(
&self,
advertiser_id: &str,
line_item_id: &str,
targeting_type: &str,
assigned_targeting_option_id: &str
) -> AdvertiserLineItemTargetingTypeAssignedTargetingOptionGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a single targeting option assigned to a line item.

Arguments

advertiserId - Required. The ID of the advertiser the line item belongs to.

lineItemId - Required. The ID of the line item the assigned targeting option belongs to.

targetingType - Required. Identifies the type of this assigned targeting option.

assignedTargetingOptionId - Required. An identifier unique to the targeting type in this line item that identifies the assigned targeting option being requested.

pub fn line_items_targeting_types_assigned_targeting_options_list(
&self,
advertiser_id: &str,
line_item_id: &str,
targeting_type: &str
) -> AdvertiserLineItemTargetingTypeAssignedTargetingOptionListCall<'a, C>

Create a builder to help you perform the following task:

Lists the targeting options assigned to a line item.

Arguments

advertiserId - Required. The ID of the advertiser the line item belongs to.

lineItemId - Required. The ID of the line item to list assigned targeting options for.

targetingType - Required. Identifies the type of assigned targeting options to list.

pub fn line_items_bulk_edit_line_item_assigned_targeting_options(
&self,
request: BulkEditLineItemAssignedTargetingOptionsRequest,
advertiser_id: &str,
line_item_id: &str
) -> AdvertiserLineItemBulkEditLineItemAssignedTargetingOptionCall<'a, C>

Create a builder to help you perform the following task:

Bulk edits targeting options under a single line item. The operation will delete the assigned targeting options provided in BulkEditLineItemAssignedTargetingOptionsRequest.delete_requests and then create the assigned targeting options provided in BulkEditLineItemAssignedTargetingOptionsRequest.create_requests .

Arguments

request - No description provided.

advertiserId - Required. The ID of the advertiser the line item belongs to.

lineItemId - Required. The ID of the line item the assigned targeting option will belong to.

pub fn line_items_bulk_list_line_item_assigned_targeting_options(
&self,
advertiser_id: &str,
line_item_id: &str
) -> AdvertiserLineItemBulkListLineItemAssignedTargetingOptionCall<'a, C>

Create a builder to help you perform the following task:

Lists assigned targeting options of a line item across targeting types.

Arguments

advertiserId - Required. The ID of the advertiser the line item belongs to.

lineItemId - Required. The ID of the line item to list assigned targeting options for.

pub fn line_items_create(
&self,
request: LineItem,
advertiser_id: &str
) -> AdvertiserLineItemCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new line item. Returns the newly created line item if successful.

Arguments

request - No description provided.

advertiserId - Output only. The unique ID of the advertiser the line item belongs to.

pub fn line_items_delete(
&self,
advertiser_id: &str,
line_item_id: &str
) -> AdvertiserLineItemDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a line item. Returns error code NOT_FOUND if the line item does not exist. The line item should be archived first, i.e. set entity_status to ENTITY_STATUS_ARCHIVED, to be able to delete it.

Arguments

advertiserId - The ID of the advertiser this line item belongs to.

lineItemId - The ID of the line item we need to fetch.

pub fn line_items_generate_default(
&self,
request: GenerateDefaultLineItemRequest,
advertiser_id: &str
) -> AdvertiserLineItemGenerateDefaultCall<'a, C>

Create a builder to help you perform the following task:

Creates a new line item with settings (including targeting) inherited from the insertion order. Returns the newly created line item if successful. There are default values based on the three fields: * The insertion order's InsertionOrderType * The insertion order's InsertionOrderAutomationType * The given line_item_type

Arguments

request - No description provided.

advertiserId - Required. The ID of the advertiser this line item belongs to.

pub fn line_items_get(
&self,
advertiser_id: &str,
line_item_id: &str
) -> AdvertiserLineItemGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a line item.

Arguments

advertiserId - Required. The ID of the advertiser this line item belongs to.

lineItemId - Required. The ID of the line item to fetch.

pub fn line_items_list(
&self,
advertiser_id: &str
) -> AdvertiserLineItemListCall<'a, C>

Create a builder to help you perform the following task:

Lists line items in an advertiser. The order is defined by the order_by parameter. If a filter by entity_status is not specified, line items with ENTITY_STATUS_ARCHIVED will not be included in the results.

Arguments

advertiserId - Required. The ID of the advertiser to list line items for.

pub fn line_items_patch(
&self,
request: LineItem,
advertiser_id: &str,
line_item_id: &str
) -> AdvertiserLineItemPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing line item. Returns the updated line item if successful.

Arguments

request - No description provided.

advertiserId - Output only. The unique ID of the advertiser the line item belongs to.

lineItemId - Output only. The unique ID of the line item. Assigned by the system.

pub fn location_lists_assigned_locations_bulk_edit(
&self,
request: BulkEditAssignedLocationsRequest,
advertiser_id: &str,
location_list_id: &str
) -> AdvertiserLocationListAssignedLocationBulkEditCall<'a, C>

Create a builder to help you perform the following task:

Bulk edits multiple assignments between locations and a single location list. The operation will delete the assigned locations provided in BulkEditAssignedLocationsRequest.deleted_assigned_locations and then create the assigned locations provided in BulkEditAssignedLocationsRequest.created_assigned_locations.

Arguments

request - No description provided.

advertiserId - Required. The ID of the DV360 advertiser to which the location list belongs.

locationListId - Required. The ID of the location list to which these assignments are assigned.

pub fn location_lists_assigned_locations_create(
&self,
request: AssignedLocation,
advertiser_id: &str,
location_list_id: &str
) -> AdvertiserLocationListAssignedLocationCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an assignment between a location and a location list.

Arguments

request - No description provided.

advertiserId - Required. The ID of the DV360 advertiser to which the location list belongs.

locationListId - Required. The ID of the location list for which the assignment will be created.

pub fn location_lists_assigned_locations_delete(
&self,
advertiser_id: &str,
location_list_id: &str,
assigned_location_id: &str
) -> AdvertiserLocationListAssignedLocationDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the assignment between a location and a location list.

Arguments

advertiserId - Required. The ID of the DV360 advertiser to which the location list belongs.

locationListId - Required. The ID of the location list to which this assignment is assigned.

assignedLocationId - Required. The ID of the assigned location to delete.

pub fn location_lists_assigned_locations_list(
&self,
advertiser_id: &str,
location_list_id: &str
) -> AdvertiserLocationListAssignedLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists locations assigned to a location list.

Arguments

advertiserId - Required. The ID of the DV360 advertiser to which the location list belongs.

locationListId - Required. The ID of the location list to which these assignments are assigned.

pub fn location_lists_create(
&self,
request: LocationList,
advertiser_id: &str
) -> AdvertiserLocationListCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new location list. Returns the newly created location list if successful.

Arguments

request - No description provided.

advertiserId - Required. The ID of the DV360 advertiser to which the location list belongs.

pub fn location_lists_get(
&self,
advertiser_id: &str,
location_list_id: &str
) -> AdvertiserLocationListGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a location list.

Arguments

advertiserId - Required. The ID of the DV360 advertiser to which the fetched location list belongs.

locationListId - Required. The ID of the location list to fetch.

pub fn location_lists_list(
&self,
advertiser_id: &str
) -> AdvertiserLocationListListCall<'a, C>

Create a builder to help you perform the following task:

Lists location lists based on a given advertiser id.

Arguments

advertiserId - Required. The ID of the DV360 advertiser to which the fetched location lists belong.

pub fn location_lists_patch(
&self,
request: LocationList,
advertiser_id: &str,
location_list_id: &str
) -> AdvertiserLocationListPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a location list. Returns the updated location list if successful.

Arguments

request - No description provided.

advertiserId - Required. The ID of the DV360 advertiser to which the location lists belongs.

locationListId - Output only. The unique ID of the location list. Assigned by the system.

pub fn manual_triggers_activate(
&self,
request: ActivateManualTriggerRequest,
advertiser_id: &str,
trigger_id: &str
) -> AdvertiserManualTriggerActivateCall<'a, C>

Create a builder to help you perform the following task:

Activates a manual trigger. Each activation of the manual trigger must be at least 5 minutes apart, otherwise an error will be returned.

Arguments

request - No description provided.

advertiserId - Required. The ID of the advertiser that the manual trigger belongs.

triggerId - Required. The ID of the manual trigger to activate.

pub fn manual_triggers_create(
&self,
request: ManualTrigger,
advertiser_id: &str
) -> AdvertiserManualTriggerCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new manual trigger. Returns the newly created manual trigger if successful.

Arguments

request - No description provided.

advertiserId - Required. Immutable. The unique ID of the advertiser that the manual trigger belongs to.

pub fn manual_triggers_deactivate(
&self,
request: DeactivateManualTriggerRequest,
advertiser_id: &str,
trigger_id: &str
) -> AdvertiserManualTriggerDeactivateCall<'a, C>

Create a builder to help you perform the following task:

Deactivates a manual trigger.

Arguments

request - No description provided.

advertiserId - Required. The ID of the advertiser that the manual trigger belongs.

triggerId - Required. The ID of the manual trigger to deactivate.

pub fn manual_triggers_get(
&self,
advertiser_id: &str,
trigger_id: &str
) -> AdvertiserManualTriggerGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a manual trigger.

Arguments

advertiserId - Required. The ID of the advertiser this manual trigger belongs to.

triggerId - Required. The ID of the manual trigger to fetch.

pub fn manual_triggers_list(
&self,
advertiser_id: &str
) -> AdvertiserManualTriggerListCall<'a, C>

Create a builder to help you perform the following task:

Lists manual triggers that are accessible to the current user for a given advertiser ID. The order is defined by the order_by parameter. A single advertiser_id is required.

Arguments

advertiserId - Required. The ID of the advertiser that the fetched manual triggers belong to.

pub fn manual_triggers_patch(
&self,
request: ManualTrigger,
advertiser_id: &str,
trigger_id: &str
) -> AdvertiserManualTriggerPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a manual trigger. Returns the updated manual trigger if successful.

Arguments

request - No description provided.

advertiserId - Required. Immutable. The unique ID of the advertiser that the manual trigger belongs to.

triggerId - Output only. The unique ID of the manual trigger.

pub fn negative_keyword_lists_negative_keywords_bulk_edit(
&self,
request: BulkEditNegativeKeywordsRequest,
advertiser_id: &str,
negative_keyword_list_id: &str
) -> AdvertiserNegativeKeywordListNegativeKeywordBulkEditCall<'a, C>

Create a builder to help you perform the following task:

Bulk edits negative keywords in a single negative keyword list. The operation will delete the negative keywords provided in BulkEditNegativeKeywordsRequest.deleted_negative_keywords and then create the negative keywords provided in BulkEditNegativeKeywordsRequest.created_negative_keywords. This operation is guaranteed to be atomic and will never result in a partial success or partial failure.

Arguments

request - No description provided.

advertiserId - Required. The ID of the DV360 advertiser to which the parent negative keyword list belongs.

negativeKeywordListId - Required. The ID of the parent negative keyword list to which the negative keywords belong.

pub fn negative_keyword_lists_negative_keywords_create(
&self,
request: NegativeKeyword,
advertiser_id: &str,
negative_keyword_list_id: &str
) -> AdvertiserNegativeKeywordListNegativeKeywordCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a negative keyword in a negative keyword list.

Arguments

request - No description provided.

advertiserId - Required. The ID of the DV360 advertiser to which the parent negative keyword list belongs.

negativeKeywordListId - Required. The ID of the parent negative keyword list in which the negative keyword will be created.

pub fn negative_keyword_lists_negative_keywords_delete(
&self,
advertiser_id: &str,
negative_keyword_list_id: &str,
keyword_value: &str
) -> AdvertiserNegativeKeywordListNegativeKeywordDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a negative keyword from a negative keyword list.

Arguments

advertiserId - Required. The ID of the DV360 advertiser to which the parent negative keyword list belongs.

negativeKeywordListId - Required. The ID of the parent negative keyword list to which the negative keyword belongs.

keywordValue - Required. The keyword value of the negative keyword to delete.

pub fn negative_keyword_lists_negative_keywords_list(
&self,
advertiser_id: &str,
negative_keyword_list_id: &str
) -> AdvertiserNegativeKeywordListNegativeKeywordListCall<'a, C>

Create a builder to help you perform the following task:

Lists negative keywords in a negative keyword list.

Arguments

advertiserId - Required. The ID of the DV360 advertiser to which the parent negative keyword list belongs.

negativeKeywordListId - Required. The ID of the parent negative keyword list to which the requested negative keywords belong.

pub fn negative_keyword_lists_negative_keywords_replace(
&self,
request: ReplaceNegativeKeywordsRequest,
advertiser_id: &str,
negative_keyword_list_id: &str
) -> AdvertiserNegativeKeywordListNegativeKeywordReplaceCall<'a, C>

Create a builder to help you perform the following task:

Replaces all negative keywords in a single negative keyword list. The operation will replace the keywords in a negative keywords with keywords provided in ReplaceNegativeKeywordsRequest.new_negative_keywords.

Arguments

request - No description provided.

advertiserId - Required. The ID of the DV360 advertiser to which the parent negative keyword list belongs.

negativeKeywordListId - Required. The ID of the parent negative keyword list to which the negative keywords belong.

pub fn negative_keyword_lists_create(
&self,
request: NegativeKeywordList,
advertiser_id: &str
) -> AdvertiserNegativeKeywordListCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new negative keyword list. Returns the newly created negative keyword list if successful.

Arguments

request - No description provided.

advertiserId - Required. The ID of the DV360 advertiser to which the negative keyword list will belong.

pub fn negative_keyword_lists_delete(
&self,
advertiser_id: &str,
negative_keyword_list_id: &str
) -> AdvertiserNegativeKeywordListDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a negative keyword list given an advertiser ID and a negative keyword list ID.

Arguments

advertiserId - Required. The ID of the DV360 advertiser to which the negative keyword list belongs.

negativeKeywordListId - Required. The ID of the negative keyword list to delete.

pub fn negative_keyword_lists_get(
&self,
advertiser_id: &str,
negative_keyword_list_id: &str
) -> AdvertiserNegativeKeywordListGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a negative keyword list given an advertiser ID and a negative keyword list ID.

Arguments

advertiserId - Required. The ID of the DV360 advertiser to which the fetched negative keyword list belongs.

negativeKeywordListId - Required. The ID of the negative keyword list to fetch.

pub fn negative_keyword_lists_list(
&self,
advertiser_id: &str
) -> AdvertiserNegativeKeywordListListCall<'a, C>

Create a builder to help you perform the following task:

Lists negative keyword lists based on a given advertiser id.

Arguments

advertiserId - Required. The ID of the DV360 advertiser to which the fetched negative keyword lists belong.

pub fn negative_keyword_lists_patch(
&self,
request: NegativeKeywordList,
advertiser_id: &str,
negative_keyword_list_id: &str
) -> AdvertiserNegativeKeywordListPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a negative keyword list. Returns the updated negative keyword list if successful.

Arguments

request - No description provided.

advertiserId - Required. The ID of the DV360 advertiser to which the negative keyword list belongs.

negativeKeywordListId - Output only. The unique ID of the negative keyword list. Assigned by the system.

pub fn targeting_types_assigned_targeting_options_create(
&self,
request: AssignedTargetingOption,
advertiser_id: &str,
targeting_type: &str
) -> AdvertiserTargetingTypeAssignedTargetingOptionCreateCall<'a, C>

Create a builder to help you perform the following task:

Assigns a targeting option to an advertiser. Returns the assigned targeting option if successful.

Arguments

request - No description provided.

advertiserId - Required. The ID of the advertiser.

targetingType - Required. Identifies the type of this assigned targeting option. Supported targeting types: * TARGETING_TYPE_CHANNEL * TARGETING_TYPE_DIGITAL_CONTENT_LABEL_EXCLUSION * TARGETING_TYPE_SENSITIVE_CATEGORY_EXCLUSION

pub fn targeting_types_assigned_targeting_options_delete(
&self,
advertiser_id: &str,
targeting_type: &str,
assigned_targeting_option_id: &str
) -> AdvertiserTargetingTypeAssignedTargetingOptionDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an assigned targeting option from an advertiser.

Arguments

advertiserId - Required. The ID of the advertiser.

targetingType - Required. Identifies the type of this assigned targeting option. Supported targeting types: * TARGETING_TYPE_CHANNEL * TARGETING_TYPE_DIGITAL_CONTENT_LABEL_EXCLUSION * TARGETING_TYPE_SENSITIVE_CATEGORY_EXCLUSION

assignedTargetingOptionId - Required. The ID of the assigned targeting option to delete.

pub fn targeting_types_assigned_targeting_options_get(
&self,
advertiser_id: &str,
targeting_type: &str,
assigned_targeting_option_id: &str
) -> AdvertiserTargetingTypeAssignedTargetingOptionGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a single targeting option assigned to an advertiser.

Arguments

advertiserId - Required. The ID of the advertiser.

targetingType - Required. Identifies the type of this assigned targeting option. Supported targeting types: * TARGETING_TYPE_CHANNEL * TARGETING_TYPE_DIGITAL_CONTENT_LABEL_EXCLUSION * TARGETING_TYPE_SENSITIVE_CATEGORY_EXCLUSION

assignedTargetingOptionId - Required. An identifier unique to the targeting type in this advertiser that identifies the assigned targeting option being requested.

pub fn targeting_types_assigned_targeting_options_list(
&self,
advertiser_id: &str,
targeting_type: &str
) -> AdvertiserTargetingTypeAssignedTargetingOptionListCall<'a, C>

Create a builder to help you perform the following task:

Lists the targeting options assigned to an advertiser.

Arguments

advertiserId - Required. The ID of the advertiser.

targetingType - Required. Identifies the type of assigned targeting options to list. Supported targeting types: * TARGETING_TYPE_CHANNEL * TARGETING_TYPE_DIGITAL_CONTENT_LABEL_EXCLUSION * TARGETING_TYPE_SENSITIVE_CATEGORY_EXCLUSION

pub fn audit(&self, advertiser_id: &str) -> AdvertiserAuditCall<'a, C>

Create a builder to help you perform the following task:

Audits an advertiser. Returns the counts of used entities per resource type under the advertiser provided. Used entities count towards their respective resource limit. See https://support.google.com/displayvideo/answer/6071450.

Arguments

advertiserId - Required. The ID of the advertiser to audit.

pub fn bulk_edit_advertiser_assigned_targeting_options(
&self,
request: BulkEditAdvertiserAssignedTargetingOptionsRequest,
advertiser_id: &str
) -> AdvertiserBulkEditAdvertiserAssignedTargetingOptionCall<'a, C>

Create a builder to help you perform the following task:

Bulk edits targeting options under a single advertiser. The operation will delete the assigned targeting options provided in BulkEditAdvertiserAssignedTargetingOptionsRequest.delete_requests and then create the assigned targeting options provided in BulkEditAdvertiserAssignedTargetingOptionsRequest.create_requests .

Arguments

request - No description provided.

advertiserId - Required. The ID of the advertiser.

pub fn bulk_list_advertiser_assigned_targeting_options(
&self,
advertiser_id: &str
) -> AdvertiserBulkListAdvertiserAssignedTargetingOptionCall<'a, C>

Create a builder to help you perform the following task:

Lists assigned targeting options of an advertiser across targeting types.

Arguments

advertiserId - Required. The ID of the advertiser the line item belongs to.

pub fn create(&self, request: Advertiser) -> AdvertiserCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new advertiser. Returns the newly created advertiser if successful. This method can take up to 180 seconds to complete.

Arguments

request - No description provided.

pub fn delete(&self, advertiser_id: &str) -> AdvertiserDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an advertiser. Deleting an advertiser will delete all of its child resources, for example, campaigns, insertion orders and line items. A deleted advertiser cannot be recovered.

Arguments

advertiserId - The ID of the advertiser we need to delete.

pub fn get(&self, advertiser_id: &str) -> AdvertiserGetCall<'a, C>

Create a builder to help you perform the following task:

Gets an advertiser.

Arguments

advertiserId - Required. The ID of the advertiser to fetch.

pub fn list(&self) -> AdvertiserListCall<'a, C>

Create a builder to help you perform the following task:

Lists advertisers that are accessible to the current user. The order is defined by the order_by parameter. A single partner_id is required. Cross-partner listing is not supported.

pub fn patch(
&self,
request: Advertiser,
advertiser_id: &str
) -> AdvertiserPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing advertiser. Returns the updated advertiser if successful.

Arguments

request - No description provided.

advertiserId - Output only. The unique ID of the advertiser. Assigned by the system.