Struct google_displayvideo1::api::AssignedTargetingOption

pub struct AssignedTargetingOption {
    pub age_range_details: Option<AgeRangeAssignedTargetingOptionDetails>,
    pub app_category_details: Option<AppCategoryAssignedTargetingOptionDetails>,
    pub app_details: Option<AppAssignedTargetingOptionDetails>,
    pub assigned_targeting_option_id: Option<String>,
    pub audience_group_details: Option<AudienceGroupAssignedTargetingOptionDetails>,
    pub authorized_seller_status_details: Option<AuthorizedSellerStatusAssignedTargetingOptionDetails>,
    pub browser_details: Option<BrowserAssignedTargetingOptionDetails>,
    pub carrier_and_isp_details: Option<CarrierAndIspAssignedTargetingOptionDetails>,
    pub category_details: Option<CategoryAssignedTargetingOptionDetails>,
    pub channel_details: Option<ChannelAssignedTargetingOptionDetails>,
    pub content_instream_position_details: Option<ContentInstreamPositionAssignedTargetingOptionDetails>,
    pub content_outstream_position_details: Option<ContentOutstreamPositionAssignedTargetingOptionDetails>,
    pub day_and_time_details: Option<DayAndTimeAssignedTargetingOptionDetails>,
    pub device_make_model_details: Option<DeviceMakeModelAssignedTargetingOptionDetails>,
    pub device_type_details: Option<DeviceTypeAssignedTargetingOptionDetails>,
    pub digital_content_label_exclusion_details: Option<DigitalContentLabelAssignedTargetingOptionDetails>,
    pub environment_details: Option<EnvironmentAssignedTargetingOptionDetails>,
    pub exchange_details: Option<ExchangeAssignedTargetingOptionDetails>,
    pub gender_details: Option<GenderAssignedTargetingOptionDetails>,
    pub geo_region_details: Option<GeoRegionAssignedTargetingOptionDetails>,
    pub household_income_details: Option<HouseholdIncomeAssignedTargetingOptionDetails>,
    pub inheritance: Option<String>,
    pub inventory_source_details: Option<InventorySourceAssignedTargetingOptionDetails>,
    pub inventory_source_group_details: Option<InventorySourceGroupAssignedTargetingOptionDetails>,
    pub keyword_details: Option<KeywordAssignedTargetingOptionDetails>,
    pub language_details: Option<LanguageAssignedTargetingOptionDetails>,
    pub name: Option<String>,
    pub negative_keyword_list_details: Option<NegativeKeywordListAssignedTargetingOptionDetails>,
    pub on_screen_position_details: Option<OnScreenPositionAssignedTargetingOptionDetails>,
    pub operating_system_details: Option<OperatingSystemAssignedTargetingOptionDetails>,
    pub parental_status_details: Option<ParentalStatusAssignedTargetingOptionDetails>,
    pub proximity_location_list_details: Option<ProximityLocationListAssignedTargetingOptionDetails>,
    pub regional_location_list_details: Option<RegionalLocationListAssignedTargetingOptionDetails>,
    pub sensitive_category_exclusion_details: Option<SensitiveCategoryAssignedTargetingOptionDetails>,
    pub sub_exchange_details: Option<SubExchangeAssignedTargetingOptionDetails>,
    pub targeting_type: Option<String>,
    pub third_party_verifier_details: Option<ThirdPartyVerifierAssignedTargetingOptionDetails>,
    pub url_details: Option<UrlAssignedTargetingOptionDetails>,
    pub user_rewarded_content_details: Option<UserRewardedContentAssignedTargetingOptionDetails>,
    pub video_player_size_details: Option<VideoPlayerSizeAssignedTargetingOptionDetails>,
    pub viewability_details: Option<ViewabilityAssignedTargetingOptionDetails>,
}

A single assigned targeting option, which defines the state of a targeting option for an entity with targeting settings.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

campaigns targeting types assigned targeting options get advertisers (response)

insertion orders targeting types assigned targeting options get advertisers (response)

line items targeting types assigned targeting options create advertisers (request|response)

line items targeting types assigned targeting options get advertisers (response)

targeting types assigned targeting options create advertisers (request|response)

targeting types assigned targeting options get advertisers (response)

targeting types assigned targeting options create partners (request|response)

targeting types assigned targeting options get partners (response)

Fields
age_range_details: Option<AgeRangeAssignedTargetingOptionDetails>

Age range details. This field will be populated when the targeting_type is TARGETING_TYPE_AGE_RANGE.

app_category_details: Option<AppCategoryAssignedTargetingOptionDetails>

App category details. This field will be populated when the targeting_type is TARGETING_TYPE_APP_CATEGORY.

app_details: Option<AppAssignedTargetingOptionDetails>

App details. This field will be populated when the targeting_type is TARGETING_TYPE_APP.

assigned_targeting_option_id: Option<String>

Output only. The unique ID of the assigned targeting option. The ID is only unique within a given resource and targeting type. It may be reused in other contexts.

audience_group_details: Option<AudienceGroupAssignedTargetingOptionDetails>

Audience targeting details. This field will be populated when the targeting_type is TARGETING_TYPE_AUDIENCE_GROUP. You can only target one audience group option per resource.

authorized_seller_status_details: Option<AuthorizedSellerStatusAssignedTargetingOptionDetails>

Authorized seller status details. This field will be populated when the targeting_type is TARGETING_TYPE_AUTHORIZED_SELLER_STATUS. You can only target one authorized seller status option per resource. If a resource doesn't have an authorized seller status option, all authorized sellers indicated as DIRECT or RESELLER in the ads.txt file are targeted by default.

browser_details: Option<BrowserAssignedTargetingOptionDetails>

Browser details. This field will be populated when the targeting_type is TARGETING_TYPE_BROWSER.

carrier_and_isp_details: Option<CarrierAndIspAssignedTargetingOptionDetails>

Carrier and ISP details. This field will be populated when the targeting_type is TARGETING_TYPE_CARRIER_AND_ISP.

category_details: Option<CategoryAssignedTargetingOptionDetails>

Category details. This field will be populated when the targeting_type is TARGETING_TYPE_CATEGORY. Targeting a category will also target its subcategories. If a category is excluded from targeting and a subcategory is included, the exclusion will take precedence.

channel_details: Option<ChannelAssignedTargetingOptionDetails>

Channel details. This field will be populated when the targeting_type is TARGETING_TYPE_CHANNEL.

content_instream_position_details: Option<ContentInstreamPositionAssignedTargetingOptionDetails>

Content instream position details. This field will be populated when the targeting_type is TARGETING_TYPE_CONTENT_INSTREAM_POSITION.

content_outstream_position_details: Option<ContentOutstreamPositionAssignedTargetingOptionDetails>

Content outstream position details. This field will be populated when the targeting_type is TARGETING_TYPE_CONTENT_OUTSTREAM_POSITION.

day_and_time_details: Option<DayAndTimeAssignedTargetingOptionDetails>

Day and time details. This field will be populated when the targeting_type is TARGETING_TYPE_DAY_AND_TIME.

device_make_model_details: Option<DeviceMakeModelAssignedTargetingOptionDetails>

Device make and model details. This field will be populated when the targeting_type is TARGETING_TYPE_DEVICE_MAKE_MODEL.

device_type_details: Option<DeviceTypeAssignedTargetingOptionDetails>

Device Type details. This field will be populated when the targeting_type is TARGETING_TYPE_DEVICE_TYPE.

digital_content_label_exclusion_details: Option<DigitalContentLabelAssignedTargetingOptionDetails>

Digital content label details. This field will be populated when the targeting_type is TARGETING_TYPE_DIGITAL_CONTENT_LABEL_EXCLUSION. Digital content labels are targeting exclusions. Advertiser level digital content label exclusions, if set, are always applied in serving (even though they aren't visible in resource settings). Resource settings can exclude content labels in addition to advertiser exclusions, but can't override them. A line item won't serve if all the digital content labels are excluded.

environment_details: Option<EnvironmentAssignedTargetingOptionDetails>

Environment details. This field will be populated when the targeting_type is TARGETING_TYPE_ENVIRONMENT.

exchange_details: Option<ExchangeAssignedTargetingOptionDetails>

Exchange details. This field will be populated when the targeting_type is TARGETING_TYPE_EXCHANGE.

gender_details: Option<GenderAssignedTargetingOptionDetails>

Gender details. This field will be populated when the targeting_type is TARGETING_TYPE_GENDER.

geo_region_details: Option<GeoRegionAssignedTargetingOptionDetails>

Geographic region details. This field will be populated when the targeting_type is TARGETING_TYPE_GEO_REGION.

household_income_details: Option<HouseholdIncomeAssignedTargetingOptionDetails>

Household income details. This field will be populated when the targeting_type is TARGETING_TYPE_HOUSEHOLD_INCOME.

inheritance: Option<String>

Output only. The inheritance status of the assigned targeting option.

inventory_source_details: Option<InventorySourceAssignedTargetingOptionDetails>

Inventory source details. This field will be populated when the targeting_type is TARGETING_TYPE_INVENTORY_SOURCE.

inventory_source_group_details: Option<InventorySourceGroupAssignedTargetingOptionDetails>

Inventory source group details. This field will be populated when the targeting_type is TARGETING_TYPE_INVENTORY_SOURCE_GROUP.

keyword_details: Option<KeywordAssignedTargetingOptionDetails>

Keyword details. This field will be populated when the targeting_type is TARGETING_TYPE_KEYWORD. A maximum of 5000 direct negative keywords can be assigned to a resource. No limit on number of positive keywords that can be assigned.

language_details: Option<LanguageAssignedTargetingOptionDetails>

Language details. This field will be populated when the targeting_type is TARGETING_TYPE_LANGUAGE.

name: Option<String>

Output only. The resource name for this assigned targeting option.

negative_keyword_list_details: Option<NegativeKeywordListAssignedTargetingOptionDetails>

Keyword details. This field will be populated when the targeting_type is TARGETING_TYPE_NEGATIVE_KEYWORD_LIST. A maximum of 4 negative keyword lists can be assigned to a resource.

on_screen_position_details: Option<OnScreenPositionAssignedTargetingOptionDetails>

On screen position details. This field will be populated when the targeting_type is TARGETING_TYPE_ON_SCREEN_POSITION.

operating_system_details: Option<OperatingSystemAssignedTargetingOptionDetails>

Operating system details. This field will be populated when the targeting_type is TARGETING_TYPE_OPERATING_SYSTEM.

parental_status_details: Option<ParentalStatusAssignedTargetingOptionDetails>

Parental status details. This field will be populated when the targeting_type is TARGETING_TYPE_PARENTAL_STATUS.

proximity_location_list_details: Option<ProximityLocationListAssignedTargetingOptionDetails>

Proximity location list details. This field will be populated when the targeting_type is TARGETING_TYPE_PROXIMITY_LOCATION_LIST.

regional_location_list_details: Option<RegionalLocationListAssignedTargetingOptionDetails>

Regional location list details. This field will be populated when the targeting_type is TARGETING_TYPE_REGIONAL_LOCATION_LIST.

sensitive_category_exclusion_details: Option<SensitiveCategoryAssignedTargetingOptionDetails>

Sensitive category details. This field will be populated when the targeting_type is TARGETING_TYPE_SENSITIVE_CATEGORY_EXCLUSION. Sensitive categories are targeting exclusions. Advertiser level sensitive category exclusions, if set, are always applied in serving (even though they aren't visible in resource settings). Resource settings can exclude sensitive categories in addition to advertiser exclusions, but can't override them.

sub_exchange_details: Option<SubExchangeAssignedTargetingOptionDetails>

Sub-exchange details. This field will be populated when the targeting_type is TARGETING_TYPE_SUB_EXCHANGE.

targeting_type: Option<String>

Output only. Identifies the type of this assigned targeting option.

third_party_verifier_details: Option<ThirdPartyVerifierAssignedTargetingOptionDetails>

Third party verification details. This field will be populated when the targeting_type is TARGETING_TYPE_THIRD_PARTY_VERIFIER.

url_details: Option<UrlAssignedTargetingOptionDetails>

URL details. This field will be populated when the targeting_type is TARGETING_TYPE_URL.

user_rewarded_content_details: Option<UserRewardedContentAssignedTargetingOptionDetails>

User rewarded content details. This field will be populated when the targeting_type is TARGETING_TYPE_USER_REWARDED_CONTENT.

video_player_size_details: Option<VideoPlayerSizeAssignedTargetingOptionDetails>

Video player size details. This field will be populated when the targeting_type is TARGETING_TYPE_VIDEO_PLAYER_SIZE.

viewability_details: Option<ViewabilityAssignedTargetingOptionDetails>

Viewability details. This field will be populated when the targeting_type is TARGETING_TYPE_VIEWABILITY. You can only target one viewability option per resource.