Struct google_displayvideo1::api::Creative

pub struct Creative {
    pub additional_dimensions: Option<Vec<Dimensions>>,
    pub advertiser_id: Option<String>,
    pub appended_tag: Option<String>,
    pub assets: Option<Vec<AssetAssociation>>,
    pub cm_placement_id: Option<String>,
    pub cm_tracking_ad: Option<CmTrackingAd>,
    pub companion_creative_ids: Option<Vec<String>>,
    pub counter_events: Option<Vec<CounterEvent>>,
    pub create_time: Option<String>,
    pub creative_attributes: Option<Vec<String>>,
    pub creative_id: Option<String>,
    pub creative_type: Option<String>,
    pub dimensions: Option<Dimensions>,
    pub display_name: Option<String>,
    pub dynamic: Option<bool>,
    pub entity_status: Option<String>,
    pub exit_events: Option<Vec<ExitEvent>>,
    pub expand_on_hover: Option<bool>,
    pub expanding_direction: Option<String>,
    pub hosting_source: Option<String>,
    pub html5_video: Option<bool>,
    pub ias_campaign_monitoring: Option<bool>,
    pub integration_code: Option<String>,
    pub js_tracker_url: Option<String>,
    pub line_item_ids: Option<Vec<String>>,
    pub media_duration: Option<String>,
    pub mp3_audio: Option<bool>,
    pub name: Option<String>,
    pub notes: Option<String>,
    pub oba_icon: Option<ObaIcon>,
    pub ogg_audio: Option<bool>,
    pub progress_offset: Option<AudioVideoOffset>,
    pub require_html5: Option<bool>,
    pub require_mraid: Option<bool>,
    pub require_ping_for_attribution: Option<bool>,
    pub review_status: Option<ReviewStatusInfo>,
    pub skip_offset: Option<AudioVideoOffset>,
    pub skippable: Option<bool>,
    pub third_party_tag: Option<String>,
    pub third_party_urls: Option<Vec<ThirdPartyUrl>>,
    pub timer_events: Option<Vec<TimerEvent>>,
    pub tracker_urls: Option<Vec<String>>,
    pub transcodes: Option<Vec<Transcode>>,
    pub universal_ad_id: Option<UniversalAdId>,
    pub update_time: Option<String>,
    pub vast_tag_url: Option<String>,
    pub vpaid: Option<bool>,
}

A single Creative.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

creatives create advertisers (request|response)

creatives get advertisers (response)

creatives patch advertisers (request|response)

Fields
additional_dimensions: Option<Vec<Dimensions>>

Additional dimensions. Applicable when creative_type is one of: * CREATIVE_TYPE_STANDARD * CREATIVE_TYPE_EXPANDABLE * CREATIVE_TYPE_NATIVE * CREATIVE_TYPE_TEMPLATED_APP_INSTALL * CREATIVE_TYPE_NATIVE_SITE_SQUARE * CREATIVE_TYPE_LIGHTBOX * CREATIVE_TYPE_NATIVE_APP_INSTALL * CREATIVE_TYPE_NATIVE_APP_INSTALL_SQUARE * CREATIVE_TYPE_PUBLISHER_HOSTED If this field is specified, width_pixels and height_pixels are both required and must be greater than or equal to 0.

advertiser_id: Option<String>

Output only. The unique ID of the advertiser the creative belongs to.

appended_tag: Option<String>

Third-party HTML tracking tag to be appended to the creative tag.

assets: Option<Vec<AssetAssociation>>

Required. Assets associated to this creative. Assets can be associated to the creative in one of following roles: * ASSET_ROLE_UNSPECIFIED * ASSET_ROLE_MAIN * ASSET_ROLE_BACKUP * ASSET_ROLE_POLITE_LOAD

cm_placement_id: Option<String>

Output only. The unique ID of the Campaign Manager 360 placement associated with the creative. This field is only applicable for creatives that are synced from Campaign Manager.

cm_tracking_ad: Option<CmTrackingAd>

The Campaign Manager 360 tracking ad associated with the creative. Optional for the following creative_type when created by an advertiser that uses both Campaign Manager 360 and third-party ad serving: * CREATIVE_TYPE_NATIVE * CREATIVE_TYPE_NATIVE_SITE_SQUARE * CREATIVE_TYPE_NATIVE_APP_INSTALL * CREATIVE_TYPE_NATIVE_APP_INSTALL_SQUARE Output only for other cases.

companion_creative_ids: Option<Vec<String>>

The IDs of companion creatives for a video creative. You can assign existing display creatives (with image or HTML5 assets) to serve surrounding the publisher's video player. Companions display around the video player while the video is playing and remain after the video has completed. Creatives contain additional dimensions can not be companion creatives. This field is only supported for following creative_type: * CREATIVE_TYPE_AUDIO * CREATIVE_TYPE_VIDEO

counter_events: Option<Vec<CounterEvent>>

Counter events for a rich media creative. Counters track the number of times that a user interacts with any part of a rich media creative in a specified way (mouse-overs, mouse-outs, clicks, taps, data loading, keyboard entries, etc.). Any event that can be captured in the creative can be recorded as a counter. Leave it empty or unset for creatives containing image assets only.

create_time: Option<String>

Output only. The timestamp when the creative was created. Assigned by the system.

creative_attributes: Option<Vec<String>>

Output only. A list of attributes of the creative that is generated by the system.

creative_id: Option<String>

Output only. The unique ID of the creative. Assigned by the system.

creative_type: Option<String>

Required. Immutable. The type of the creative.

dimensions: Option<Dimensions>

Required. Primary dimensions of the creative. Applicable to all creative types. The value of width_pixels and height_pixels defaults to 0 when creative_type is one of: * CREATIVE_TYPE_VIDEO * CREATIVE_TYPE_TEMPLATED_APP_INSTALL_INTERSTITIAL * CREATIVE_TYPE_AUDIO * CREATIVE_TYPE_NATIVE_VIDEO * CREATIVE_TYPE_TEMPLATED_APP_INSTALL_VIDEO

display_name: Option<String>

Required. The display name of the creative. Must be UTF-8 encoded with a maximum size of 240 bytes.

dynamic: Option<bool>

Output only. Indicates whether the creative is dynamic.

entity_status: Option<String>

Required. Controls whether or not the creative can serve. Accepted values are: * ENTITY_STATUS_ACTIVE * ENTITY_STATUS_ARCHIVED * ENTITY_STATUS_PAUSED

exit_events: Option<Vec<ExitEvent>>

Required. Exit events for this creative. An exit (also known as a click tag) is any area in your creative that someone can click or tap to open an advertiser's landing page. Every creative must include at least one exit. You can add an exit to your creative in any of the following ways: * Use Google Web Designer's tap area. * Define a JavaScript variable called "clickTag". * Use the Enabler (Enabler.exit()) to track exits in rich media formats.

expand_on_hover: Option<bool>

Optional. Indicates the creative will automatically expand on hover. Optional and only valid for third-party expandable creatives. Third-party expandable creatives are creatives with following hosting source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_EXPANDABLE

expanding_direction: Option<String>

Optional. Specifies the expanding direction of the creative. Required and only valid for third-party expandable creatives. Third-party expandable creatives are creatives with following hosting source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_EXPANDABLE

hosting_source: Option<String>

Required. Indicates where the creative is hosted.

html5_video: Option<bool>

Output only. Indicates the third-party VAST tag creative requires HTML5 Video support. Output only and only valid for third-party VAST tag creatives. Third-party VAST tag creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_VIDEO

ias_campaign_monitoring: Option<bool>

Indicates whether Integral Ad Science (IAS) campaign monitoring is enabled. To enable this for the creative, make sure the Advertiser.creative_config.ias_client_id has been set to your IAS client ID.

integration_code: Option<String>

ID information used to link this creative to an external system. Must be UTF-8 encoded with a length of no more than 10,000 characters.

js_tracker_url: Option<String>

JavaScript measurement URL from supported third-party verification providers (ComScore, DoubleVerify, IAS, Moat). HTML script tags are not supported. This field is only supported in following creative_type: * CREATIVE_TYPE_NATIVE * CREATIVE_TYPE_NATIVE_SITE_SQUARE * CREATIVE_TYPE_NATIVE_APP_INSTALL * CREATIVE_TYPE_NATIVE_APP_INSTALL_SQUARE * CREATIVE_TYPE_NATIVE_VIDEO

line_item_ids: Option<Vec<String>>

Output only. The IDs of the line items this creative is associated with. To associate a creative to a line item, use LineItem.creative_ids instead.

media_duration: Option<String>

Output only. Media duration of the creative. Applicable when creative_type is one of: * CREATIVE_TYPE_VIDEO * CREATIVE_TYPE_AUDIO * CREATIVE_TYPE_NATIVE_VIDEO * CREATIVE_TYPE_PUBLISHER_HOSTED

mp3_audio: Option<bool>

Output only. Indicates the third-party audio creative supports MP3. Output only and only valid for third-party audio creatives. Third-party audio creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_AUDIO

name: Option<String>

Output only. The resource name of the creative.

notes: Option<String>

User notes for this creative. Must be UTF-8 encoded with a length of no more than 20,000 characters.

oba_icon: Option<ObaIcon>

Specifies the OBA icon for a video creative. This field is only supported in following creative_type: * CREATIVE_TYPE_VIDEO

ogg_audio: Option<bool>

Output only. Indicates the third-party audio creative supports OGG. Output only and only valid for third-party audio creatives. Third-party audio creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_AUDIO

progress_offset: Option<AudioVideoOffset>

Amount of time to play the video before counting a view. This field is required when skippable is true. This field is only supported for the following creative_type: * CREATIVE_TYPE_VIDEO

require_html5: Option<bool>

Optional. Indicates that the creative relies on HTML5 to render properly. Optional and only valid for third-party tag creatives. Third-party tag creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_STANDARD * CREATIVE_TYPE_EXPANDABLE

require_mraid: Option<bool>

Optional. Indicates that the creative requires MRAID (Mobile Rich Media Ad Interface Definitions system). Set this if the creative relies on mobile gestures for interactivity, such as swiping or tapping. Optional and only valid for third-party tag creatives. Third-party tag creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_STANDARD * CREATIVE_TYPE_EXPANDABLE

require_ping_for_attribution: Option<bool>

Optional. Indicates that the creative will wait for a return ping for attribution. Only valid when using a Campaign Manager 360 tracking ad with a third-party ad server parameter and the ${DC_DBM_TOKEN} macro. Optional and only valid for third-party tag creatives or third-party VAST tag creatives. Third-party tag creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_STANDARD * CREATIVE_TYPE_EXPANDABLE Third-party VAST tag creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_AUDIO * CREATIVE_TYPE_VIDEO

review_status: Option<ReviewStatusInfo>

Output only. The current status of the creative review process.

skip_offset: Option<AudioVideoOffset>

Amount of time to play the video before the skip button appears. This field is required when skippable is true. This field is only supported for the following creative_type: * CREATIVE_TYPE_VIDEO

skippable: Option<bool>

Whether the user can choose to skip a video creative. This field is only supported for the following creative_type: * CREATIVE_TYPE_VIDEO

third_party_tag: Option<String>

Optional. The original third-party tag used for the creative. Required and only valid for third-party tag creatives. Third-party tag creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_STANDARD * CREATIVE_TYPE_EXPANDABLE

third_party_urls: Option<Vec<ThirdPartyUrl>>

Tracking URLs from third parties to track interactions with a video creative. This field is only supported for the following creative_type: * CREATIVE_TYPE_AUDIO * CREATIVE_TYPE_VIDEO * CREATIVE_TYPE_NATIVE_VIDEO

timer_events: Option<Vec<TimerEvent>>

Timer custom events for a rich media creative. Timers track the time during which a user views and interacts with a specified part of a rich media creative. A creative can have multiple timer events, each timed independently. Leave it empty or unset for creatives containing image assets only.

tracker_urls: Option<Vec<String>>

Tracking URLs for analytics providers or third-party ad technology vendors. The URLs must start with https (except on inventory that doesn't require SSL compliance). If using macros in your URL, use only macros supported by Display & Video 360. Standard URLs only, no IMG or SCRIPT tags. This field is only supported in following creative_type: * CREATIVE_TYPE_NATIVE * CREATIVE_TYPE_NATIVE_SITE_SQUARE * CREATIVE_TYPE_NATIVE_APP_INSTALL * CREATIVE_TYPE_NATIVE_APP_INSTALL_SQUARE * CREATIVE_TYPE_NATIVE_VIDEO

transcodes: Option<Vec<Transcode>>

Output only. Audio/Video transcodes. Display & Video 360 transcodes the main asset into a number of alternative versions that use different file formats or have different properties (resolution, audio bit rate, and video bit rate), each designed for specific video players or bandwidths. These transcodes give a publisher's system more options to choose from for each impression on your video and ensures that the appropriate file serves based on the viewer’s connection and screen size. This field is only supported in following creative_type: * CREATIVE_TYPE_VIDEO * CREATIVE_TYPE_NATIVE_VIDEO * CREATIVE_TYPE_AUDIO

universal_ad_id: Option<UniversalAdId>

Optional. An optional creative identifier provided by a registry that is unique across all platforms. Universal Ad ID is part of the VAST 4.0 standard. It can be modified after the creative is created. This field is only supported for the following creative_type: * CREATIVE_TYPE_VIDEO

update_time: Option<String>

Output only. The timestamp when the creative was last updated. Assigned by the system.

vast_tag_url: Option<String>

Optional. The URL of the VAST tag for a third-party VAST tag creative. Required and only valid for third-party VAST tag creatives. Third-party VAST tag creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_AUDIO * CREATIVE_TYPE_VIDEO

vpaid: Option<bool>

Output only. Indicates the third-party VAST tag creative requires VPAID (Digital Video Player-Ad Interface). Output only and only valid for third-party VAST tag creatives. Third-party VAST tag creatives are creatives with following hosting_source: * HOSTING_SOURCE_THIRD_PARTY combined with following creative_type: * CREATIVE_TYPE_VIDEO