Struct google_displayvideo1::api::InventorySourceGroupMethods

pub struct InventorySourceGroupMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on inventorySourceGroup resources. It is not used directly, but through the DisplayVideo hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_displayvideo1 as displayvideo1;
 
use std::default::Default;
use oauth2;
use displayvideo1::DisplayVideo;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = DisplayVideo::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `assigned_inventory_sources_bulk_edit(...)`, `assigned_inventory_sources_create(...)`, `assigned_inventory_sources_delete(...)`, `assigned_inventory_sources_list(...)`, `create(...)`, `delete(...)`, `get(...)`, `list(...)` and `patch(...)`
// to build up your call.
let rb = hub.inventory_source_groups();

Implementations

impl<'a, C> InventorySourceGroupMethods<'a, C>

pub fn assigned_inventory_sources_bulk_edit(
&self,
request: BulkEditAssignedInventorySourcesRequest,
inventory_source_group_id: &str
) -> InventorySourceGroupAssignedInventorySourceBulkEditCall<'a, C>

Create a builder to help you perform the following task:

Bulk edits multiple assignments between inventory sources and a single inventory source group. The operation will delete the assigned inventory sources provided in BulkEditAssignedInventorySourcesRequest.deleted_assigned_inventory_sources and then create the assigned inventory sources provided in BulkEditAssignedInventorySourcesRequest.created_assigned_inventory_sources.

Arguments

request - No description provided.

inventorySourceGroupId - Required. The ID of the inventory source group to which the assignments are assigned.

pub fn assigned_inventory_sources_create(
&self,
request: AssignedInventorySource,
inventory_source_group_id: &str
) -> InventorySourceGroupAssignedInventorySourceCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an assignment between an inventory source and an inventory source group.

Arguments

request - No description provided.

inventorySourceGroupId - Required. The ID of the inventory source group to which the assignment will be assigned.

pub fn assigned_inventory_sources_delete(
&self,
inventory_source_group_id: &str,
assigned_inventory_source_id: &str
) -> InventorySourceGroupAssignedInventorySourceDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the assignment between an inventory source and an inventory source group.

Arguments

inventorySourceGroupId - Required. The ID of the inventory source group to which this assignment is assigned.

assignedInventorySourceId - Required. The ID of the assigned inventory source to delete.

pub fn assigned_inventory_sources_list(
&self,
inventory_source_group_id: &str
) -> InventorySourceGroupAssignedInventorySourceListCall<'a, C>

Create a builder to help you perform the following task:

Lists inventory sources assigned to an inventory source group.

Arguments

inventorySourceGroupId - Required. The ID of the inventory source group to which these assignments are assigned.

pub fn create(
&self,
request: InventorySourceGroup
) -> InventorySourceGroupCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new inventory source group. Returns the newly created inventory source group if successful.

Arguments

request - No description provided.

pub fn delete(
&self,
inventory_source_group_id: &str
) -> InventorySourceGroupDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an inventory source group.

Arguments

inventorySourceGroupId - Required. The ID of the inventory source group to delete.

pub fn get(
&self,
inventory_source_group_id: &str
) -> InventorySourceGroupGetCall<'a, C>

Create a builder to help you perform the following task:

Gets an inventory source group.

Arguments

inventorySourceGroupId - Required. The ID of the inventory source group to fetch.

pub fn list(&self) -> InventorySourceGroupListCall<'a, C>

Create a builder to help you perform the following task:

Lists inventory source groups that are accessible to the current user. The order is defined by the order_by parameter.

pub fn patch(
&self,
request: InventorySourceGroup,
inventory_source_group_id: &str
) -> InventorySourceGroupPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an inventory source group. Returns the updated inventory source group if successful.

Arguments

request - No description provided.

inventorySourceGroupId - Output only. The unique ID of the inventory source group. Assigned by the system.