Struct google_displayvideo1::api::LineItem

pub struct LineItem {
    pub advertiser_id: Option<String>,
    pub bid_strategy: Option<BiddingStrategy>,
    pub budget: Option<LineItemBudget>,
    pub campaign_id: Option<String>,
    pub conversion_counting: Option<ConversionCountingConfig>,
    pub creative_ids: Option<Vec<String>>,
    pub display_name: Option<String>,
    pub entity_status: Option<String>,
    pub flight: Option<LineItemFlight>,
    pub frequency_cap: Option<FrequencyCap>,
    pub insertion_order_id: Option<String>,
    pub integration_details: Option<IntegrationDetails>,
    pub inventory_source_ids: Option<Vec<String>>,
    pub line_item_id: Option<String>,
    pub line_item_type: Option<String>,
    pub mobile_app: Option<MobileApp>,
    pub name: Option<String>,
    pub pacing: Option<Pacing>,
    pub partner_costs: Option<Vec<PartnerCost>>,
    pub partner_revenue_model: Option<PartnerRevenueModel>,
    pub targeting_expansion: Option<TargetingExpansionConfig>,
    pub update_time: Option<String>,
    pub warning_messages: Option<Vec<String>>,
}

A single line item.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

line items create advertisers (request|response)

line items generate default advertisers (response)

line items get advertisers (response)

line items patch advertisers (request|response)

Fields
advertiser_id: Option<String>

Output only. The unique ID of the advertiser the line item belongs to.

bid_strategy: Option<BiddingStrategy>

Required. The bidding strategy of the line item.

budget: Option<LineItemBudget>

Required. The budget allocation setting of the line item.

campaign_id: Option<String>

Output only. The unique ID of the campaign that the line item belongs to.

conversion_counting: Option<ConversionCountingConfig>

The conversion tracking setting of the line item.

creative_ids: Option<Vec<String>>

The IDs of the creatives associated with the line item.

display_name: Option<String>

Required. The display name of the line item. Must be UTF-8 encoded with a maximum size of 240 bytes.

entity_status: Option<String>

Required. Controls whether or not the line item can spend its budget and bid on inventory. * For CreateLineItem method, only ENTITY_STATUS_DRAFT is allowed. To activate a line item, use UpdateLineItem method and update the status to ENTITY_STATUS_ACTIVE after creation. * A line item cannot be changed back to ENTITY_STATUS_DRAFT status from any other status. * If the line item's parent insertion order is not active, the line item can't spend its budget even if its own status is ENTITY_STATUS_ACTIVE.

flight: Option<LineItemFlight>

Required. The start and end time of the line item's flight.

frequency_cap: Option<FrequencyCap>

Required. The frequency capping setting of the line item.

insertion_order_id: Option<String>

Required. Immutable. The unique ID of the insertion order that the line item belongs to.

integration_details: Option<IntegrationDetails>

Integration details of the line item.

inventory_source_ids: Option<Vec<String>>

The IDs of the private inventory sources assigned to the line item.

line_item_id: Option<String>

Output only. The unique ID of the line item. Assigned by the system.

line_item_type: Option<String>

Required. Immutable. The type of the line item.

mobile_app: Option<MobileApp>

The mobile app promoted by the line item. This is applicable only when line_item_type is either LINE_ITEM_TYPE_DISPLAY_MOBILE_APP_INSTALL or LINE_ITEM_TYPE_VIDEO_MOBILE_APP_INSTALL.

name: Option<String>

Output only. The resource name of the line item.

pacing: Option<Pacing>

Required. The budget spending speed setting of the line item.

partner_costs: Option<Vec<PartnerCost>>

The partner costs associated with the line item. If absent or empty in CreateLineItem method, the newly created line item will inherit partner costs from its parent insertion order.

partner_revenue_model: Option<PartnerRevenueModel>

Required. The partner revenue model setting of the line item.

targeting_expansion: Option<TargetingExpansionConfig>

The targeting expansion settings of the line item. This config is only applicable when eligible audience list targeting is assigned to the line item.

update_time: Option<String>

Output only. The timestamp when the line item was last updated. Assigned by the system.

warning_messages: Option<Vec<String>>

Output only. The warning messages generated by the line item. These warnings do not block saving the line item, but some may block the line item from running.