Struct google_displayvideo1::api::PartnerMethods

pub struct PartnerMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on partner resources. It is not used directly, but through the DisplayVideo hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_displayvideo1 as displayvideo1;
 
use std::default::Default;
use oauth2;
use displayvideo1::DisplayVideo;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = DisplayVideo::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `bulk_edit_partner_assigned_targeting_options(...)`, `channels_create(...)`, `channels_get(...)`, `channels_list(...)`, `channels_patch(...)`, `channels_sites_bulk_edit(...)`, `channels_sites_create(...)`, `channels_sites_delete(...)`, `channels_sites_list(...)`, `channels_sites_replace(...)`, `get(...)`, `list(...)`, `targeting_types_assigned_targeting_options_create(...)`, `targeting_types_assigned_targeting_options_delete(...)`, `targeting_types_assigned_targeting_options_get(...)` and `targeting_types_assigned_targeting_options_list(...)`
// to build up your call.
let rb = hub.partners();

Implementations

impl<'a, C> PartnerMethods<'a, C>

pub fn channels_sites_bulk_edit(
&self,
request: BulkEditSitesRequest,
partner_id: &str,
channel_id: &str
) -> PartnerChannelSiteBulkEditCall<'a, C>

Create a builder to help you perform the following task:

Bulk edits sites under a single channel. The operation will delete the sites provided in BulkEditSitesRequest.deleted_sites and then create the sites provided in BulkEditSitesRequest.created_sites.

Arguments

request - No description provided.

partnerId - The ID of the partner that owns the parent channel.

channelId - Required. The ID of the parent channel to which the sites belong.

pub fn channels_sites_create(
&self,
request: Site,
partner_id: &str,
channel_id: &str
) -> PartnerChannelSiteCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a site in a channel.

Arguments

request - No description provided.

partnerId - The ID of the partner that owns the parent channel.

channelId - Required. The ID of the parent channel in which the site will be created.

pub fn channels_sites_delete(
&self,
partner_id: &str,
channel_id: &str,
url_or_app_id: &str
) -> PartnerChannelSiteDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a site from a channel.

Arguments

partnerId - The ID of the partner that owns the parent channel.

channelId - Required. The ID of the parent channel to which the site belongs.

urlOrAppId - Required. The URL or app ID of the site to delete.

pub fn channels_sites_list(
&self,
partner_id: &str,
channel_id: &str
) -> PartnerChannelSiteListCall<'a, C>

Create a builder to help you perform the following task:

Lists sites in a channel.

Arguments

partnerId - The ID of the partner that owns the parent channel.

channelId - Required. The ID of the parent channel to which the requested sites belong.

pub fn channels_sites_replace(
&self,
request: ReplaceSitesRequest,
partner_id: &str,
channel_id: &str
) -> PartnerChannelSiteReplaceCall<'a, C>

Create a builder to help you perform the following task:

Replaces all of the sites under a single channel. The operation will replace the sites under a channel with the sites provided in ReplaceSitesRequest.new_sites.

Arguments

request - No description provided.

partnerId - The ID of the partner that owns the parent channel.

channelId - Required. The ID of the parent channel whose sites will be replaced.

pub fn channels_create(
&self,
request: Channel,
partner_id: &str
) -> PartnerChannelCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new channel. Returns the newly created channel if successful.

Arguments

request - No description provided.

partnerId - The ID of the partner that owns the created channel.

pub fn channels_get(
&self,
partner_id: &str,
channel_id: &str
) -> PartnerChannelGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a channel for a partner or advertiser.

Arguments

partnerId - The ID of the partner that owns the fetched channel.

channelId - Required. The ID of the channel to fetch.

pub fn channels_list(&self, partner_id: &str) -> PartnerChannelListCall<'a, C>

Create a builder to help you perform the following task:

Lists channels for a partner or advertiser.

Arguments

partnerId - The ID of the partner that owns the channels.

pub fn channels_patch(
&self,
request: Channel,
partner_id: &str,
channel_id: &str
) -> PartnerChannelPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a channel. Returns the updated channel if successful.

Arguments

request - No description provided.

partnerId - The ID of the partner that owns the created channel.

channelId - Output only. The unique ID of the channel. Assigned by the system.

pub fn targeting_types_assigned_targeting_options_create(
&self,
request: AssignedTargetingOption,
partner_id: &str,
targeting_type: &str
) -> PartnerTargetingTypeAssignedTargetingOptionCreateCall<'a, C>

Create a builder to help you perform the following task:

Assigns a targeting option to a partner. Returns the assigned targeting option if successful.

Arguments

request - No description provided.

partnerId - Required. The ID of the partner.

targetingType - Required. Identifies the type of this assigned targeting option. Supported targeting types: * TARGETING_TYPE_CHANNEL

pub fn targeting_types_assigned_targeting_options_delete(
&self,
partner_id: &str,
targeting_type: &str,
assigned_targeting_option_id: &str
) -> PartnerTargetingTypeAssignedTargetingOptionDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an assigned targeting option from a partner.

Arguments

partnerId - Required. The ID of the partner.

targetingType - Required. Identifies the type of this assigned targeting option. Supported targeting types: * TARGETING_TYPE_CHANNEL

assignedTargetingOptionId - Required. The ID of the assigned targeting option to delete.

pub fn targeting_types_assigned_targeting_options_get(
&self,
partner_id: &str,
targeting_type: &str,
assigned_targeting_option_id: &str
) -> PartnerTargetingTypeAssignedTargetingOptionGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a single targeting option assigned to a partner.

Arguments

partnerId - Required. The ID of the partner.

targetingType - Required. Identifies the type of this assigned targeting option. Supported targeting types: * TARGETING_TYPE_CHANNEL

assignedTargetingOptionId - Required. An identifier unique to the targeting type in this partner that identifies the assigned targeting option being requested.

pub fn targeting_types_assigned_targeting_options_list(
&self,
partner_id: &str,
targeting_type: &str
) -> PartnerTargetingTypeAssignedTargetingOptionListCall<'a, C>

Create a builder to help you perform the following task:

Lists the targeting options assigned to a partner.

Arguments

partnerId - Required. The ID of the partner.

targetingType - Required. Identifies the type of assigned targeting options to list. Supported targeting types: * TARGETING_TYPE_CHANNEL

pub fn bulk_edit_partner_assigned_targeting_options(
&self,
request: BulkEditPartnerAssignedTargetingOptionsRequest,
partner_id: &str
) -> PartnerBulkEditPartnerAssignedTargetingOptionCall<'a, C>

Create a builder to help you perform the following task:

Bulk edits targeting options under a single partner. The operation will delete the assigned targeting options provided in BulkEditPartnerAssignedTargetingOptionsRequest.deleteRequests and then create the assigned targeting options provided in BulkEditPartnerAssignedTargetingOptionsRequest.createRequests .

Arguments

request - No description provided.

partnerId - Required. The ID of the partner.

pub fn get(&self, partner_id: &str) -> PartnerGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a partner.

Arguments

partnerId - Required. The ID of the partner to fetch.

pub fn list(&self) -> PartnerListCall<'a, C>

Create a builder to help you perform the following task:

Lists partners that are accessible to the current user. The order is defined by the order_by parameter.